### 2025-05-17 #8225 @editorialbot
[PRE REVIEW]: asyncmd: A python library to orchestrate complex molecular dynamics simulation campaigns on high performance computing systems
**Submitting author:** @hejung (<a href="http://orcid.org/0000-0002-2159-0391">Hendrik Jung</a>)
**Repository:** https://github.com/bio-phys/asyncmd
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.3.3
