### 2025-05-21 f01e84b
fix dois for joss publication
