+++ joss_paper/paper.bib
@@ -4,7 +4,7 @@ and Wickberg, Tim",
 editor="Klus{\'a}{\v{c}}ek, Dalibor
 and Corbal{\'a}n, Julita
 and Rodrigo, Gonzalo P.",
-title="Architecture of the Slurm Workload Manager",
+title="Architecture of the Slurm Workload Manager",
 booktitle="Job Scheduling Strategies for Parallel Processing",
 year="2023",
 publisher="Springer Nature Switzerland",
@@ -12,6 +12,7 @@ address="Cham",
 pages="3--23",
 abstract="Slurm is an open source, fault-tolerant, and highly scalable workload manager used on many of the world's supercomputers and computer clusters. As a cluster workload manager, Slurm has three key functions. First, it allocates exclusive and/or non-exclusive access to resources for some duration of time. Second, it provides a framework for starting, executing, and monitoring work on the allocated resources. Finally, it arbitrates contention for resources by managing queues of pending work and enforcing administrative policies. This paper describes the current design and capabilities of Slurm.",
 isbn="978-3-031-43943-8"
+doi="10.1007/978-3-031-43943-8_1"
 }
 @article{gromacs2020,
     author = {Páll, Szilárd and Zhmurov, Artem and Bauer, Paul and Abraham, Mark and Lundborg, Magnus and Gray, Alan and Hess, Berk and Lindahl, Erik},
@@ -36,7 +37,7 @@ volume = {32},
 number = {10},
 pages = {2319-2327},
 keywords = {molecular dynamics simulations, analysis, proteins, object-oriented design, software, membrane systems, Python programming language},
-doi = {https://doi.org/10.1002/jcc.21787},
+doi = {10.1002/jcc.21787},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jcc.21787},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/jcc.21787},
 abstract = {Abstract MDAnalysis is an object-oriented library for structural and temporal analysis of molecular dynamics (MD) simulation trajectories and individual protein structures. It is written in the Python language with some performance-critical code in C. It uses the powerful NumPy package to expose trajectory data as fast and efficient NumPy arrays. It has been tested on systems of millions of particles. Many common file formats of simulation packages including CHARMM, Gromacs, Amber, and NAMD and the Protein Data Bank format can be read and written. Atoms can be selected with a syntax similar to CHARMM's powerful selection commands. MDAnalysis enables both novice and experienced programmers to rapidly write their own analytical tools and access data stored in trajectories in an easily accessible manner that facilitates interactive explorative analysis. MDAnalysis has been tested on and works for most Unix-based platforms such as Linux and Mac OS X. It is freely available under the GNU General Public License from http://mdanalysis.googlecode.com. © 2011 Wiley Periodicals, Inc. J Comput Chem 2011},
