### 2025-05-21 cd489db
fix dois for joss publication, attempt 2
