+++ joss_paper/paper.bib
@@ -11,7 +11,7 @@ publisher="Springer Nature Switzerland",
 address="Cham",
 pages="3--23",
 abstract="Slurm is an open source, fault-tolerant, and highly scalable workload manager used on many of the world's supercomputers and computer clusters. As a cluster workload manager, Slurm has three key functions. First, it allocates exclusive and/or non-exclusive access to resources for some duration of time. Second, it provides a framework for starting, executing, and monitoring work on the allocated resources. Finally, it arbitrates contention for resources by managing queues of pending work and enforcing administrative policies. This paper describes the current design and capabilities of Slurm.",
-isbn="978-3-031-43943-8"
+isbn="978-3-031-43943-8",
 doi="10.1007/978-3-031-43943-8_1"
 }
 @article{gromacs2020,
