### 2025-05-27 04035d4
Improved consistency check for mdp options when performin enegry minimization with gromacs engines
