### 2025-05-27 2276211
Improved wording in changelog
