### 2025-05-28 ee715f3
removed some obsoltete comments from gromacs engine and updated mdengine baseclass definition
