### 2025-05-28 395c408
Some tests for gromacs engine
