### 2025-05-29 2273b0c
More tests for GmxEngine
