+++ tests/gromacs/test_mdengine.py
@@ -14,8 +14,17 @@
 # along with asyncmd. If not, see <https://www.gnu.org/licenses/>.
 import pytest
 import logging
+import shutil
+
+import numpy as np
 
 from asyncmd.gromacs import GmxEngine, MDP
+from asyncmd import Trajectory
+
+
+# Decorator for test that need gmx grompp and/or gmx mdrun
+needs_gmx_install = pytest.mark.skipif(shutil.which("gmx") is None,
+                                       reason="Need gromacs (gmx) executable.")
 
 
 class Test_GmxEngine:
@@ -25,6 +34,7 @@ class Test_GmxEngine:
         self.ndx = "tests/test_data/gromcas/index.ndx"
         self.top = "tests/test_data/gromacs/topol_amber99sbildn.top"
         self.mdp_md_compressed_out = MDP("tests/test_data/gromacs/md_compressed_out.mdp")
+        self.mdp_md_full_prec_out = MDP("tests/test_data/gromacs/md_full_prec_out.mdp")
 
     @pytest.mark.parametrize("integrator", ["steep", "cg", "l-bfgs"])
     def test_mpd_check_minimize(self, integrator, monkeypatch):
@@ -39,6 +49,10 @@ class Test_GmxEngine:
         self.mdp_md_compressed_out["integrator"] = integrator
         with pytest.raises(ValueError):
             engine.mdp = self.mdp_md_compressed_out
+        # this should work
+        engine.output_traj_type = "trr"
+        self.mdp_md_full_prec_out["integrator"] = integrator
+        engine.mdp = self.mdp_md_full_prec_out
 
     def test_mdp_check_nsteps(self, caplog, monkeypatch):
         # init an engine so we can use its mdconfig property (which does the checks)
@@ -80,3 +94,50 @@ class Test_GmxEngine:
                 engine = GmxEngine(mdconfig=None,
                                    gro_file=self.gro,
                                    top_file=self.top)
+
+    @pytest.mark.slow
+    @needs_gmx_install
+    @pytest.mark.parametrize("starting_conf",
+                             [None,
+                              Trajectory("tests/test_data/trajectory/ala_traj.trr",
+                                         "tests/test_data/trajectory/ala.tpr")
+                              ]
+                             )
+    @pytest.mark.asyncio
+    async def test_run_MD_compressed_out(self, tmpdir, starting_conf):
+        engine = GmxEngine(mdconfig=self.mdp_md_compressed_out,
+                           gro_file=self.gro,
+                           top_file=self.top)
+        await engine.prepare(starting_configuration=starting_conf,
+                             workdir=tmpdir,
+                             deffnm="test")
+        nsteps = 10
+        traj = await engine.run(nsteps=nsteps)
+        # some basic checks
+        assert len(traj) == engine.nstout / nsteps + 1
+        assert engine.steps_done == nsteps
+        assert np.isclose(engine.time_done, nsteps * engine.dt)
+
+    @pytest.mark.slow
+    @needs_gmx_install
+    @pytest.mark.parametrize("starting_conf",
+                             [None,
+                              Trajectory("tests/test_data/trajectory/ala_traj.trr",
+                                         "tests/test_data/trajectory/ala.tpr")
+                              ]
+                             )
+    @pytest.mark.asyncio
+    async def test_run_MD_full_prec_out(self, tmpdir, starting_conf):
+        engine = GmxEngine(mdconfig=self.mdp_md_full_prec_out,
+                           gro_file=self.gro,
+                           top_file=self.top,
+                           output_traj_type="trr")
+        await engine.prepare(starting_configuration=starting_conf,
+                             workdir=tmpdir,
+                             deffnm="test")
[…]
