### 2025-05-29 50b9728
more GmxEngine tests
