### 2025-05-30 1f1677d
A bit of cleanup for gromacs engine
