### 2025-05-30 #8321 @editorialbot
[REVIEW]: asyncmd: A python library to orchestrate complex molecular dynamics simulation campaigns on high performance computing systems
**Submitting author:** @hejung (<a href="http://orcid.org/0000-0002-2159-0391">Hendrik Jung</a>)
**Repository:** https://github.com/bio-phys/asyncmd
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.4.1
