### 2025-06-04 1b2f146
Correction for slurm tests
