+++ tests/trajectory/test_propagate.py
@@ -13,17 +13,200 @@
 # You should have received a copy of the GNU General Public License
 # along with asyncmd. If not, see <https://www.gnu.org/licenses/>.
 
-import os
+import logging
 
 import numpy as np
 import pytest
 
+from asyncmd import Trajectory
 from asyncmd import gromacs as asyncgmx
 from asyncmd import trajectory as asynctraj
 
 
-def condition_function(traj):
-    return np.array([False, False, False, True, False])
+def dummy_condition_func(traj, return_len, true_on_frames):
+    """
+    Construct a condition function return for ConditionalTrajectoryPropagators.
+
+    Actually construct a fixed return value for the function according to given
+    specifications.
+
+    Parameters
+    ----------
+    return_len : int
+        The number of frames/values the function returns.
+    true_on_frames : list[int]
+        Which of these frames should be true (all others will be false).
+        Zero-indexed!
+    """
+    # NOTE: we add the return_len and true_on_frames (which fully specify the
+    # return value) as call_kwargs, such that the cached values will always
+    # match the specified return (since call_kwargs is part of the function
+    # hash which we use to cache results)
+    ret = np.full((return_len,), False)
+    for frame in true_on_frames:
+        ret[frame] = True
+    return ret
+
+
+class Test_ConditionalPropagator:
+    def setup_method(self):
+        self.empty_mdp = asyncgmx.MDP("tests/test_data/gromacs/empty.mdp")
+        self.cond_wrap_1frame_true1 = asynctraj.PyTrajectoryFunctionWrapper(
+                                            function=dummy_condition_func,
+                                            call_kwargs={"return_len": 1,
+                                                         "true_on_frames": [0],
+                                                         }
+                                                                            )
+        self.cond_wrap_1frame_never_true = asynctraj.PyTrajectoryFunctionWrapper(
+                                            function=dummy_condition_func,
+                                            call_kwargs={"return_len": 1,
+                                                         "true_on_frames": [],
+                                                         }
+                                                                              )
+
+    @pytest.mark.parametrize(["max_frames", "max_steps", "beauty_max_steps"],
+                             [(None, None, float("inf")),  # no max steps/frames
+                              (200, 10, 10),  # max_steps takes precedence
+                              (10, None, 10),  # nstout=1, max_steps=max_frames
+                              (None, 10, 10),
+                              ])
+    def test_init_max_steps_vs_max_frames(self, monkeypatch, max_frames,
+                                          max_steps, beauty_max_steps):
+        with monkeypatch.context() as m:
+            # monkeypatch to make sure nstout will be equal to 1
+            m.setattr("asyncmd.trajectory.propagate.nstout_from_mdconfig",
+                      lambda mdconfig, output_traj_type: 1)
+            propa = asynctraj.ConditionalTrajectoryPropagator(
+                        conditions=[self.cond_wrap_1frame_true1,
+                                    self.cond_wrap_1frame_never_true],
+                        engine_cls=asyncgmx.GmxEngine,
+                        engine_kwargs={"mdconfig": self.empty_mdp},
+                        walltime_per_part=0.01,
+                        max_steps=max_steps,
+                        max_frames=max_frames,
+                                                              )
+        assert propa.max_steps == beauty_max_steps
+
+    def test_conditions_setter_not_all_coro(self, monkeypatch, caplog):
+        with monkeypatch.context() as m:
+            m.setattr("asyncmd.trajectory.propagate.nstout_from_mdconfig",
+                      lambda mdconfig, output_traj_type: 1)
+            propa = asynctraj.ConditionalTrajectoryPropagator(
+                        conditions=[self.cond_wrap_1frame_true1],
+                        engine_cls=asyncgmx.GmxEngine,
[…]
