### 2025-06-16 bfd96ff
Improved handling of conditions in ConditionalTrajectoryPropagator; refactor: FlagChangeList and TypedFlagChangeList moved to tools.py
