+++ tests/test_tools.py
@@ -1,0 +1,262 @@
+# This file is part of asyncmd.
+#
+# asyncmd is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# asyncmd is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with asyncmd. If not, see <https://www.gnu.org/licenses/>.
+import pytest
+import numpy as np
+
+
+from asyncmd.tools import FlagChangeList, TypedFlagChangeList
+
+
+class Test_FlagChangeList:
+    @pytest.mark.parametrize("no_iter_data",
+                             [1,  # ints are not iterable
+                              1.,  # floats are also not iterable
+                              ]
+                             )
+    def test_init_errs(self, no_iter_data):
+        with pytest.raises(TypeError):
+            _ = FlagChangeList(data=no_iter_data)
+
+    @pytest.mark.parametrize(["test_data", "data_len"],
+                             [(["test", "1", "2", "3"], 4),
+                              (["test", 1, 2, 3], 4),
+                              ([1., 2., 3., 4.], 4),
+                              ([1, 2, 3, 4], 4),
+                              ([1], 1),
+                              ([], 0),
+                              ]
+                             )
+    def test_len_getitem(self, test_data, data_len):
+        flag_list = FlagChangeList(data=test_data)
+        # check that length is correct
+        assert len(flag_list) == data_len
+        # and check that all items are what we expect
+        for idx, value in enumerate(flag_list):
+            assert value == test_data[idx]
+
+    @pytest.mark.parametrize("test_data",
+                             [["test", "1", "2", "3"],
+                              ["test", 1, 2, 3],
+                              [1., 2., 3., 4.],
+                              [1, 2, 3, 4],
+                              [1],
+                              ]
+                             )
+    def test_changed_setitem_delitem_insert(self, test_data):
+        flag_list = FlagChangeList(data=test_data)
+        assert not flag_list.changed
+        # get an item and check that everything is still good
+        _ = flag_list[0]
+        assert not flag_list.changed
+        # modify and see that we do set changed=True
+        flag_list[0] = "1234"
+        assert flag_list.changed
+        # reininit to get a new list with changed=False
+        flag_list = FlagChangeList(data=test_data)
+        assert not flag_list.changed  # as it should be
+        # now delete an item and check again
+        del flag_list[0]
+        assert flag_list.changed
+        # again reinit, this time to test insert
+        flag_list = FlagChangeList(data=test_data)
+        assert not flag_list.changed
+        obj_to_insert = object()
+        # get a random index to insert at
+        if len(flag_list) > 0:
+            idx = np.random.randint(low=0, high=len(flag_list))
+        else:
+            idx = 0
+        flag_list.insert(idx, obj_to_insert)
+        assert flag_list[idx] is obj_to_insert
+
+    @pytest.mark.parametrize("to_add",
+                             [[2, 3, 4],
+                              (2, 3, 4),
+                              FlagChangeList([2, 3, 4]),
+                              "234",
+                              ])
+    def test_add(self, to_add):
+        # we always start with a FlagChangeList with a 1
+        beauty = [1] + list(to_add)
+        flag_list = FlagChangeList(data=[1])
+        flag_list_added = flag_list + to_add
+        # make sure we created a new object
+        assert flag_list is not flag_list_added
+        # make sure the new object knows it has not been changed since creation
[…]
