+++ tests/conftest.py
@@ -15,7 +15,6 @@
 import pytest
 
 
-
 def pytest_addoption(parser):
     parser.addoption(
         "--runslow", action="store_true", default=False, help="run slow tests"
