+++ pytest.ini
@@ -1,3 +1,4 @@
 [pytest]
 asyncio_mode = strict
+asyncio_default_test_loop_scope = session
 addopts = -ra
