### 2025-06-19 2df65f6
more tests for tools.py
