### 2025-06-19 055c4fb
Some more tests for MDConfig
