### 2025-06-20 c8e3faa
move pytest marker for tests that need gromacs to conftest.py
