### 2025-06-20 4ac1374
modernize test fixtures
