### 2025-06-20 67436e8
modernize more test fixtures
