### 2025-06-20 8beed0b
Some tests for InPartsTrajectoryPropagator
