+++ tests/trajectory/test_propagate.py
@@ -12,15 +12,21 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with asyncmd. If not, see <https://www.gnu.org/licenses/>.
-
+import pytest
 import logging
+import os
 
 import numpy as np
-import pytest
 
 from asyncmd import Trajectory
 from asyncmd import gromacs as asyncgmx
-from asyncmd import trajectory as asynctraj
+from asyncmd.trajectory import (PyTrajectoryFunctionWrapper,
+                                InPartsTrajectoryPropagator,
+                                ConditionalTrajectoryPropagator,
+                                )
+
+# marker for tests that need gromacs installed
+from conftest import needs_gmx_install
 
 
 def dummy_condition_func(traj, return_len, true_on_frames):
@@ -48,21 +54,84 @@ def dummy_condition_func(traj, return_len, true_on_frames):
     return ret
 
 
+class Test_InPartsTrajectoryPropagator:
+    def setup_method(self):
+        self.empty_mdp = asyncgmx.MDP("tests/test_data/gromacs/empty.mdp")
+        self.xtc_mdp = asyncgmx.MDP("tests/test_data/gromacs/md_compressed_out.mdp")
+        self.ala_gro = "tests/test_data/gromacs/conf.gro"
+        self.ala_top = "tests/test_data/gromacs/topol_amber99sbildn.top"
+
+    @pytest.mark.slow
+    @needs_gmx_install
+    @pytest.mark.parametrize(["n_steps", "continuation_n_steps"],
+                             [(0, 0), (10, 10),
+                              (0, 10), (0, 200),
+                              (10, 0), (200, 0),
+                              (10, 100),
+                              (300, 100),
+                              ])
+    @pytest.mark.parametrize("starting_conf",
+                             [Trajectory("tests/test_data/trajectory/ala_traj.trr",
+                                         "tests/test_data/trajectory/ala.tpr")
+                              ])
+    @pytest.mark.asyncio
+    async def test_run_continue_remove_with_gromacs(
+                self, tmp_path, starting_conf, n_steps, continuation_n_steps,
+                                                    ):
+        propa = InPartsTrajectoryPropagator(
+                        n_steps=n_steps,
+                        engine_cls=asyncgmx.GmxEngine,
+                        engine_kwargs={"mdconfig": self.xtc_mdp,
+                                       "gro_file": self.ala_gro,
+                                       "top_file": self.ala_top,
+                                       },
+                        walltime_per_part=1 / (60 * 60 * 100),  # 0.01 s
+                                            )
+        traj = await propa.propagate_and_concatenate(
+                                starting_configuration=starting_conf,
+                                workdir=tmp_path,
+                                deffnm="test",
+                                tra_out=os.path.join(tmp_path, "out_traj.xtc"),
+                                )
+        if n_steps == 0:
+            assert traj is None
+        else:
+            assert len(traj) == ((n_steps / self.xtc_mdp["nstxout-compressed"])
+                                 + 1)
+        # reset n_steps
+        propa.n_steps = continuation_n_steps
+        # and do a continuation
+        traj_cont = await propa.propagate_and_concatenate(
+                            starting_configuration=starting_conf,
+                            workdir=tmp_path,
+                            deffnm="test",
+                            tra_out=os.path.join(tmp_path, "out_traj2.xtc"),
+                            continuation=True,
+                            )
+        if continuation_n_steps == 0:
+            assert traj_cont is None
+        else:
+            assert len(traj_cont) == ((continuation_n_steps
+                                       / self.xtc_mdp["nstxout-compressed"]
+                                       )
+                                      + 1)
+
[…]
