+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `SlurmGmxEngine` (and `GmxEngine`) now expose the `mdrun_time_conversion_factor` to enable users to control the expected time it takes to set up the environment inside the slurm job. Both engines also have improved consistency checks for mdp options when performing energy minimization.
 - `MDEngine`, `GmxEngine`, and `SlurmGmxEngine`: removed unused `running` property
+- Some tests for `trajectory.propagate` and `gromacs.mdengine` modules
 
 ## [0.3.3] - 2025-05-06
 
