### 2025-06-23 e706eb1
More tests for propagate.py
