+++ tests/trajectory/test_propagate.py
@@ -17,6 +17,8 @@ import logging
 import os
 
 import numpy as np
+import MDAnalysis as mda
+from MDAnalysis.lib.distances import calc_dihedrals
 
 from asyncmd import Trajectory
 from asyncmd import gromacs as asyncgmx
@@ -52,6 +54,100 @@ def dummy_condition_func(traj, return_len, true_on_frames):
     for frame in true_on_frames:
         ret[frame] = True
     return ret
+
+
+def ala_alpha_R(traj, skip=1):
+    """
+    Calculate alpha_R state function for capped alaninie dipeptide.
+
+    The alpha_R state is defined in the space of the two dihedral angles psi
+    and phi, for a configuration to belong to the state:
+        phi: -pi < phi < 0
+        psi: -50 degree < psi < 30 degree
+
+    Parameters
+    ----------
+    traj : asyncmd.Trajectory
+        The trajectory for which the state function is calculated.
+    skip : int, optional
+        stride for trajectory iteration, by default 1
+
+    Returns
+    -------
+    numpy.ndarray, shape=(n_frames,)
+        Array with boolean values for every configuration on the trajectory
+        indicating if a configuration falls into the state or not.
+    """
+    u = mda.Universe(traj.structure_file, *traj.trajectory_files)
+    psi_ag = u.select_atoms("resname ALA and name N")  # idx 6
+    psi_ag += u.select_atoms("resname ALA and name CA")  # idx 8
+    psi_ag += u.select_atoms("resname ALA and name C")  # idx 14
+    psi_ag += u.select_atoms("resname NME and name N")  # idx 16
+    phi_ag = u.select_atoms("resname ACE and name C")  # idx 4
+    phi_ag += u.select_atoms("resname ALA and name N")  # idx 6
+    phi_ag += u.select_atoms("resname ALA and name CA")  # idx 8
+    phi_ag += u.select_atoms("resname ALA and name C")  # idx 14
+    # empty arrays to fill
+    state = np.full((len(u.trajectory[::skip]),), False, dtype=bool)
+    phi = np.empty((len(u.trajectory[::skip]),), dtype=np.float64)
+    psi = np.empty((len(u.trajectory[::skip]),), dtype=np.float64)
+    for f, ts in enumerate(u.trajectory[::skip]):
+        phi[f] = calc_dihedrals(*(at.position for at in phi_ag), box=ts.dimensions)
+        psi[f] = calc_dihedrals(*(at.position for at in psi_ag), box=ts.dimensions)
+    # make sure MDAnalysis closes the underlying trajectory files directly
+    u.trajectory.close()
+    # phi: -pi -> 0
+    # psi: > -50 but smaller 30 degree
+    deg = 180/np.pi
+    state[(phi <= 0) & (-50/deg <= psi) & (psi <= 30/deg)] = True
+    return state
+
+
+def ala_C7_eq(traj, skip=1):
+    """
+    Calculate C7_eq state function for capped alanine dipeptide.
+
+    The C7_eq state is defined in the space of the two dihedral angles psi
+    and phi, for a configuration to belong to the state:
+        phi: -pi < phi < 0
+        psi: 120 degree < psi < 200 degree
+
+    Parameters
+    ----------
+    traj : asyncmd.Trajectory
+        The trajectory for which the state function is calculated.
+    skip : int, optional
+        stride for trajectory iteration, by default 1
+
+    Returns
+    -------
+    numpy.ndarray, shape=(n_frames,)
+        Array with boolean values for every configuration on the trajectory
+        indicating if a configuration falls into the state or not.
+    """
[…]
