### 2025-06-23 bio-phys/asyncmd#36 @corettialessandro
Problem reading topology files
Hi @hejung,

I installed the package both via PyPi and directly from repository and everything seems fine in both cases.
Documentation also seems fine.

Then when I run tests with pytest I get the attached output.

[pytest.log](https://github.com/user-attachments/files/20868697/pytest.log)

which seems to be related somehow to the way TPR files are read.

I tried with the three cases mentioned, namely:
- `pip install .\[tests\]`
- `pip install .\[tests-all\]`
- `pip install -e .\[dev\]`

with basically same results (log with `pip install -e .\[dev\]` was slightly different, let me know if that can help).

The same kind of problem seems to happen when I run the first example notebook GmxEngine.ipynb: At cell 18 I get the attached output:

[notebook_cell18.log](https://github.com/user-attachments/files/20868866/notebook_cell18.log)

this time complaining about the TRR files.

Is there maybe some problem with the MDAnalysis version or do you know what could be happening?
Am I maybe doing something stupid (sorry if that's the case)?

Thanks!
