### 2025-06-23 #8321 @corettialessandro
I am having trouble running tests with pytest and one example, I opened an issue here: https://github.com/bio-phys/asyncmd/issues/36
