### 2025-06-23 bio-phys/asyncmd#37 @braniii
[JOSS] Review
Hello @hejung,

sorry for the delay with the review process. I try to finish the review by tomorrow evening. Until then, I will regularly add to the list.

This issue will contain the review of the JOSS submission, https://github.com/openjournals/joss-reviews/issues/8321. Once you've completed the tasks, feel free to tick them of. In case there is something unclear, feel free to reach out.

For testing I am using `uv`:
```bash
uv venv --python 3.12
uv pip install -e .[dev]
```

#### Issues:
- [ ] Please upload the documentation and link it to the repo and README. I guess the easiest solution is using an CI together with github pages.
- [ ] Please add a note in the README that for running the tests git lfs is needed. Otherwise it results in 204 failing tests.
- [ ] It would be very valuable if you could include the example ipynbs into the documentation. This should be easy to achieve.
