### 2025-06-23 bio-phys/asyncmd#36 @braniii
@corettialessandro I got the same issue (see #37 ). To clone the test data git lfs is needed. Installing it and running the following command (within the cloned repo) fixed it for me
```bash
git lfs install
git lfs pull
```
