### 2025-06-24 dc89746
update redme to make it more clear that git-lfs is needed to checkout example and test input files
