### 2025-06-24 bio-phys/asyncmd#37 @hejung
Hi!

Thanks for the review!

Unfortunately I am away from tomorrow on and will probably not get to do much before Tuesday. I will see what I get done today in terms of uploading the documentation and including the notebooks directly into it. So the good news is there is no hurry about getting the review done before Monday evening, since that will probably keep me busy for today :)
