### 2025-06-24 bio-phys/asyncmd#36 @hejung
Hi!

Some of the binary input files for the examples and the tests are stored using git-lfs (mostly trajectories/input configurations). For now, I tried to make that more clear in the README (in #38 ).
An other option would be to just store the files without git-lfs in the repository, since they are not that large and they will (hopefully) not change often that should not increase the repository size by much.

Let me know if you think the readme update is enough!
