+++ docs/source/api.rst
@@ -1,343 +1,3 @@
-Overview for users
-==================
-
-trajectory
-**********
-The :py:mod:`asyncmd.trajectory` module contains a
-:py:class:`asyncmd.Trajectory` class which is the return object for all MD
-engines. These objects enable easy access to a number properties of the
-underlying trajectory, like the length in frames or time, the intergration step
-and many more. Note :py:class:`asyncmd.Trajectory` are unique objects in the
-sense that every combination of underlying ``trajectory_files`` will give you the
-same object back even if you instantiate it multiple times, i.e. ``is`` will be
-``True`` for the two objects (in addition to ``==`` beeing ``True``).
-Also note that it is possible to pickle and unpickle :py:class:`asyncmd.Trajectory`
-objects. You can even change the filepath of the underlying trajectories, i.e.
-copy/move them to another location (consider also moving the npz cache files)
-and still unpickle to get a working :py:class:`asyncmd.Trajectory` object as
-long as the relative path between your python workdir and the trajectory files
-does not change. Or you can change the workdir of the python interpreter as long
-as the trajectory files remain at the same location in the filesystem.
-
-..
-   TODO: reference to the dev section where we explain the two hidden funcs
-   to forget traj objects
-
-There are also a number of ``TrajectoryFunctionWrapper`` classes which
-can be used to wrapp (python) functions or arbitrary executables for easy
-asyncronous application on :py:class:`asyncmd.Trajectory`, either submitted
-via slurm or ran locally. The benefit of the wrapped functions is that the
-calculated values will be cached automatically. The caching is even persistent
-over multiple reloads and invocations of the python interpreter. To this end
-the default caching mechanism creates hidden numpy npz files for every
-:py:class:`asyncmd.Trajectory` (named after the trajectory) in which the values
-are stored. Other caching mechanism are an in-memory cache and the option to
-store all cached values in a :py:class:`h5py.File` or :py:class:`h5py.Group`.
-You can set the default caching mechanism for all
-:py:class:`asyncmd.Trajectory` centrally via
-:py:func:`asyncmd.config.set_default_trajectory_cache_type` or overwrite it for
-each :py:class:`asyncmd.Trajectory` at init by passing ``cache_type``. See also
-:py:func:`asyncmd.config.register_h5py_cache` to register the h5py cache.
-
-.. py:currentmodule:: asyncmd.trajectory.convert
-
-It also contains a number of classes to extract frames from
-:py:class:`asyncmd.Trajectory` objects in the module
-:py:mod:`asyncmd.trajectory.convert`:
-
-  - :py:class:`NoModificationFrameExtractor`
-
-  - :py:class:`InvertedVelocitiesFrameExtractor`
-
-  - :py:class:`RandomVelocitiesFrameExtractor`
-
-Note that implementing your own ``FrameExtractor`` with a custom modification
-is as easy as subclassing the abstract base class :py:class:`FrameExtractor`
-and overwriting its :py:meth:`FrameExtractor.apply_modification` method.
-
-The :py:mod:`asyncmd.trajectory.convert` module furthermore contains a class to
-concatenate :py:class:`asyncmd.Trajectory` segments, the
-:py:class:`TrajectoryConcatenator`. It can be used to concatenate lists of
-trajectory segments in any order (and possibly with inverted momenta) by
-passing a list of trajectory segments and a list of tuples (slices) that
-specify the frames to use in the concatenated output trajectory. Note that this
[…]
