+++ docs/source/index.rst
@@ -14,6 +14,8 @@ molecular dynamics simulations using pythons **async/await** synthax.
    :maxdepth: 3
    :caption: Contents:
    
+   user_docs
+   dev_docs
    api
 
 

[…]
