### 2025-06-24 3b8a5f6
get autodoc to work on RTD
