### 2025-06-24 e05ff39
try again to get autodoc to work
