+++ docs/source/conf.py
@@ -12,7 +12,10 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../../src/'))
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__),
+                                                '../../src/')
+                                   )
+                )
 
 
 # -- Project information -----------------------------------------------------
