### 2025-06-29 #8321 @braniii
Hi @espottesmith, I just wanted to let you know that I've finished the review and created a list of comments in the issue linked above. Once this has been addressed, I can check off any remaining open issues from my review checklist.
