### 2025-07-02 f62aba6
WIP: include example notebooks into documentation
