+++ docs/source/index.md
@@ -1,0 +1,70 @@
+# Welcome to asyncmd's documentation!
+
+`asyncmd` is is a library to write **concurrent** code to run and analyze
+molecular dynamics simulations using pythons **async/await** synthax.
+
+## Get started
+
+```{toctree}
+:maxdepth: 2
+:caption: The Basics
+
+usage/installation
+```
+
+## User guide
+
+**TODO: Some words on the user guide...**
+
+```{toctree}
+:maxdepth: 2
+:caption: User guide
+
+user_docs
+dev_docs
+```
+
+## Example notebooks
+
+```{toctree}
+:maxdepth: 1
+:caption: Example notebooks
+
+Overview <examples_link/README>
+GmxEngine <examples_link/01_engines/gromacs/GmxEngine>
+SlurmGmxEngine <examples_link/01_engines/gromacs/SlurmGmxEngine>
+PyTrajectoryFunctionWrapper <examples_link/02_TrajectoryFunctionWrappers/PyTrajectoryFunctionWrapper>
+SlurmTrajectoryFunctionWrapper <examples_link/02_TrajectoryFunctionWrappers/SlurmTrajectoryFunctionWrapper>
+InPartsTrajectoryPropagator <examples_link/>
+```
+
+## Community guide
+
+**TODO: Some words here**
+
+```{toctree}
+:maxdepth: 2
+:caption: Community
+
+support
+contributing/index
+```
+
+----------------
+
+## Changelog
+
+```{toctree}
+:maxdepth: 2
+
+include_changelog
+```
+
+## Indices and tables
+
+```{toctree}
+:maxdepth: 1
+
+modindex
+genindex
+```
