+++ docs/source/index.rst
@@ -1,27 +1,0 @@
-.. asyncmd documentation master file, created by
-   sphinx-quickstart on Sun Apr 24 02:14:01 2022.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to asyncmd's documentation!
-===================================
-
-`asyncmd` is is a library to write **concurrent** code to run and analyze
-molecular dynamics simulations using pythons **async/await** synthax.
-
-
-.. toctree::
-   :maxdepth: 3
-   :caption: Contents:
-   
-   user_docs
-   dev_docs
-   api
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
