+++ pyproject.toml
@@ -36,7 +36,7 @@ classifiers = [
 ]
 
 [project.optional-dependencies]
-docs = ["sphinx"]
+docs = ["sphinx", "myst-nb"]
 tests = ["pytest", "pytest-asyncio"]
 tests-all = ["asyncmd[tests]", "h5py", "coverage", "pytest-cov"]
 dev = ["asyncmd[docs,tests-all]"]
