### 2025-07-03 bio-phys/asyncmd#36 @corettialessandro
@hejung I would write a line about that also in the section that is related to the tests or add something in the installation from repository section where you mention the tests using pytest (e.g. "Please note that you need to have git-lfs (an open source git extension) setup to get all input files needed to run **the tests** and the notebooks in the examples folder").
