### 2025-07-03 d7e1c44
Change documentation theme; streamline README and move most of it to the documentation.
