+++ examples/README.md
@@ -1,5 +1,20 @@
 # `asyncmd` example jupyter notebooks
 
+```{toctree}
+:maxdepth: 1
+:caption: All notebooks
+
+GmxEngine <01_engines/gromacs/GmxEngine>
+SlurmGmxEngine <01_engines/gromacs/SlurmGmxEngine>
+PyTrajectoryFunctionWrapper <02_TrajectoryFunctionWrappers/PyTrajectoryFunctionWrapper>
+SlurmTrajectoryFunctionWrapper <02_TrajectoryFunctionWrappers/SlurmTrajectoryFunctionWrapper>
+InPartsTrajectoryPropagator <03_trajectory_propagation_and_subtrajectory_extraction/InPartsTrajectoryPropagator>
+ConditionalTrajectoryPropagator <03_trajectory_propagation_and_subtrajectory_extraction/ConditionalTrajectoryPropagator>
+FrameExtractors <03_trajectory_propagation_and_subtrajectory_extraction/FrameExtractors>
+WeightedEnsemble <04_application_examples/WeightedEnsemble>
+SlurmProcess <05_developer_topics/slurm/SlurmProcess>
+```
+
 Here you find example jupyter notebooks for various topics and tasks that `asyncmd` is suited for. If you have never used `asyncmd` you can just follow the numbering and start with `01_engines` to get an overview of the molecular dynamics engines and then proceed in the numbered order. If you are searching for something specific the folder names are hopefully descriptive enough, such that you know where to look. In any case below is a list of all example notebooks and a summary of their content.
 
 ## `01_engines`
@@ -9,27 +24,27 @@ How to use the molecular dynamics engines in `asyncmd`. The subfolders are for t
 ### `gromacs`
 
 - {doc}`GmxEngine <01_engines/gromacs/GmxEngine>`: Run gromacs on your local machine using `asyncmd`, useful mostly for learning, testing and development.
-- `SlurmGmxEngine.ipynb`: Run gromacs on your favorite HPC system via the SLURM queuing system using `asyncmd`, has the same API as the local engine, useful for production.
+- {doc}`SlurmGmxEngine <01_engines/gromacs/SlurmGmxEngine>`: Run gromacs on your favorite HPC system via the SLURM queuing system using `asyncmd`, has the same API as the local engine, useful for production.
 
 ## `02_TrajectoryFunctionWrappers`
 
-- `PyTrajectoryFunctionWrapper.ipynb`: Turn python functions acting on `asyncmd.Trajectory` objects into async functions and automatically cache the results for repeated applications. Computation is done locally in threads.
+- {doc}`PyTrajectoryFunctionWrapper <02_TrajectoryFunctionWrappers/PyTrajectoryFunctionWrapper>`: Turn python functions acting on `asyncmd.Trajectory` objects into async functions and automatically cache the results for repeated applications. Computation is done locally in threads.
 
-- `SlurmTrajectoryFunctionWrapper.ipynb`: Turn executables acting on molecular dynamics trajectories (represented as `asyncmd.Trajectory` objects) into async functions and automatically cache the results for repeated applications. Computation is submited via the SLURM queuing system.
+- {doc}`SlurmTrajectoryFunctionWrapper <02_TrajectoryFunctionWrappers/SlurmTrajectoryFunctionWrapper>`: Turn executables acting on molecular dynamics trajectories (represented as `asyncmd.Trajectory` objects) into async functions and automatically cache the results for repeated applications. Computation is submited via the SLURM queuing system.
 
 ## `03_trajectory_propagation_and_subtrajectory_extraction`
 
-- `InPartsTrajectoryPropagator.ipynb`: Run your simulation in parts of short(er) walltime for a given total number of integration steps. Useful when using slurm to make full use of backfilling and/or to run simulations that run longer than then time limit.
[…]
