### 2025-07-03 2f1d1b2
Add community section to docs (get support and contributing guidelines)
