### 2025-07-04 2f778d6
Documentation on extending asyncmd, Trajectory objects, TrajectoryFunctionWrappers, and value caching. Some minor docstring changes to make autodoc/sphinx happy.
