### 2025-07-04 e950cc0
Documentation for trajectory.convert, trajectory.propagate, gromacs and config submodules. Also some docstring changes to make autodoc/sphinx happy
