### 2025-07-04 bio-phys/asyncmd#36 @corettialessandro
Hi @hejung, so I run all the examples on my workstation and everything looks great. The Vienna Scientific Cluster has been down for some time now, so I could not test the slurm implementation, but I do not want to delay further the review process. I'll try again next week and if there is anything that comes to my mind I will just open another issue unrelated to joss.

Congratulations, it is a very nice piece of software and the example notebooks are extremely instructive. One optional thing to address is the following: I had compiled gromacs with thread_mpi, which creates issue when run with the default value of the option `mdrun_extra_args="-ntomp 2"` (probably due to a wrong setting of an environment variable on my side). It is a bit tedious to change that to `mdrun_extra_args="-nt 2"` manually in every instance where it appears, so it could be nice for the users to set it once and for all at the beginning of the notebook, maybe just after where you check for the `gmx` executable. Of course it is not a big deal so I'll leave that to you if you think it is appropriate.

Last thing about the paper, I think that a section with the state of the art would be beneficial, as suggested by @braniii. Maybe some qualitative comparison with OpenPathSampling will also be beneficial for the community.
Once this last point is addressed I will check the last tick on the checklist and the review will be completed.

Of course let me know if you have questions.
