+++ docs/source/api.rst
@@ -1,8 +1,0 @@
-API (Hierachical module layout plan)
-====================================
-
-.. autosummary::
-   :recursive:
-   :toctree: generated
-
-   asyncmd
