+++ docs/source/module_layout_autosummary.rst
@@ -1,0 +1,5 @@
+.. autosummary::
+   :recursive:
+   :toctree: generated
+
+   asyncmd
