### 2025-07-04 f43082f
Implement JOSS review (part 1) (#38)
Expand and improve documentation, deploy on read the docs (PR to address parts of #37).
