+++ docs/source/conf.py
@@ -10,19 +10,24 @@
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
-# import os
-# import sys
-# sys.path.insert(0, os.path.abspath('.'))
+import os
+import sys
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__),
+                                                '../../src/')
+                                   )
+                )
 
 
 # -- Project information -----------------------------------------------------
 
+import importlib.metadata
+
 project = 'asyncmd'
-copyright = '2022, Hendrik Jung'
+copyright = '2022-now, Hendrik Jung'
 author = 'Hendrik Jung'
 
 # The full version, including alpha/beta/rc tags
-release = '0.1.0'
+version = release = importlib.metadata.version("asyncmd")
 
 
 # -- General configuration ---------------------------------------------------
@@ -33,7 +38,8 @@ release = '0.1.0'
 extensions = [
     'sphinx.ext.autosummary',
     'sphinx.ext.autodoc',
-    'sphinx.ext.napoleon'
+    'sphinx.ext.napoleon',
+    'myst_nb',
         ]
 
 # Add any paths that contain templates here, relative to this directory.
@@ -50,7 +56,36 @@ exclude_patterns = []
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'alabaster'
+html_theme = 'sphinx_book_theme'
+
+html_theme_options = {
+    "path_to_docs": "docs/source",
+    "repository_url": "https://github.com/bio-phys/asyncmd.git",
+    "repository_branch": "main",
+
+    "use_edit_page_button": True,
+    "use_source_button": True,
+    "use_repository_button": True,
+    "use_download_button": True,
+    #"use_sidenotes": True,
+
+    "show_toc_level": 2,  # show up to one sub-heading on the right sidebar
+
+    # links with icons in left sidebar
+    "icon_links": [
+        {
+            "name": "GitHub",
+            "url": "https://github.com/bio-phys/asyncmd.git",
+            "icon": "fa-brands fa-github",
+        },
+        {
+            "name": "PyPI",
+            "url": "https://pypi.org/project/asyncmd/",
+            "icon": "https://img.shields.io/pypi/dm/asyncmd",
+            "type": "url",
+        },
+    ],
+}
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
@@ -67,3 +102,6 @@ autodoc_default_options = {
                     # document members (default = False)
                     "members": False,
                            }
+
+# -- Options for MyST-NB
+nb_execution_mode = "off"  # render the notebooks as they are in the repository
