+++ docs/source/community/support.md
@@ -1,0 +1,5 @@
+# Get support
+
+For questions related to asyncmd or to report problems with asyncmd please open a [Github issue]. Feature requests and other ideas are also welcome there!
+
+[Github issue]: https://github.com/bio-phys/asyncmd/issues
