### 2025-07-04 bio-phys/asyncmd#37 @hejung
@braniii Thanks again for the review! These are all very good points and I agree that implementing them is a good idea.

I went ahead and checked all the documentation and readme boxes after merging #38. The documentation should now be reachable https://asyncmd.readthedocs.io/en/latest/ . Please feel free to uncheck any of them again or let me know if you think they are not sufficiently addressed!

Will now implement the rest of the points, should hopefully all be done by Monday.
