### 2025-07-05 08ce024
Add jupyterlab as dev dependency
