### 2025-07-06 5040316
Notebook updates (#39)
* Add jupyterlab as dev dependency

* Update all example notebooks: define useful variables (e.g. engine_kwargs and mdrun_extra_args) to be used throughout the notebooks, remove unused imports, suppress unecessary warnings, improve spelling and structure. WEightedEnsemble notebook: Change the sampled transition to high barrier along psi to make it more obvious that WE concentrates the sampling, also add a paragraph explaining this and on why we might get lucky and observe a transition along phi in the EQ (but possibly not in the WE).
