### 2025-07-07 #8321 @hejung
@espottesmith Sorry for not writing here earlier! Most of the code issues are already addressed, I am currently finishing up the missing points from https://github.com/bio-phys/asyncmd/issues/37 and we are adding a state of the field section to the paper. I will ping you here once everything is completed.
