### 2025-07-07 c0e7102
Add pylint configuration and github workflow
