+++ pyproject.toml
@@ -10,6 +10,7 @@ dependencies = ["aiofiles",
                 "numpy",
                 "scipy",
                 ]
+# if you change requires-python also change py-version for pylint below!
 requires-python = ">=3.10"
 authors = [{ name = "Hendrik Jung", email = "hendrik.jung@biophys.mpg.de"}]
 maintainers = [{ name = "Hendrik Jung", email = "hendrik.jung@biophys.mpg.de"}]
@@ -41,6 +42,7 @@ tests = ["pytest", "pytest-asyncio"]
 tests-all = ["asyncmd[tests]", "h5py", "coverage", "pytest-cov"]
 dev = ["asyncmd[docs,tests-all]",
        "jupyterlab", "ipywidgets", "tqdm",  # needed for example notebooks
+       "pylint",
        ]
 
 [project.urls]
@@ -50,3 +52,35 @@ Issues = "https://github.com/bio-phys/asyncmd/issues"
 
 [tool.setuptools.packages.find]
 where = ["src"]
+
+# pylint configuration
+[tool.pylint.main]
+# Return non-zero exit code if any of these messages/categories are detected,
+# even if score is above --fail-under value. Syntax same as enable. Messages
+# specified are enabled, while categories only check already-enabled messages.
+fail-on = ["E"]
+
+# Specify a score threshold under which the program will exit with error.
+fail-under = 8.2
+
+# List of plugins (as comma separated values of python module names) to load,
+# usually to register additional checkers.
+load-plugins = ["pylint.extensions.bad_builtin",
+                "pylint.extensions.broad_try_clause",
+                "pylint.extensions.check_elif",
+                "pylint.extensions.code_style",
+                "pylint.extensions.comparison_placement",
+                "pylint.extensions.consider_refactoring_into_while_condition",
+                "pylint.extensions.dict_init_mutate",
+                "pylint.extensions.docparams",
+                "pylint.extensions.for_any_all",
+                "pylint.extensions.overlapping_exceptions",
+                "pylint.extensions.redefined_loop_name",
+                "pylint.extensions.redefined_variable_type",
+                "pylint.extensions.set_membership",
+                "pylint.extensions.typing",
+                ]
+
+# Minimum Python version to use for version dependent checks. Will default to the
+# version used to run pylint.
+py-version = "3.10"
