+++ .github/workflows/lint-with-pylint.yml
@@ -1,0 +1,27 @@
+name: Lint with pylint
+
+on:
+  pull_request:
+    types: [opened, synchronize, reopened, ready_for_review]
+    branches: ["main"]
+
+jobs:
+  pylint:
+    name: pylint
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v4
+      with:
+        lfs: true
+    - name: Set up Python
+      uses: actions/setup-python@v5
+      with:
+        python-version: "3.x"
+    - name: Install asyncmd and pylint
+    # install asyncmd so pylint can find the module
+      run: |
+       python -m pip install -e .
+       python -m pip install pylint
+    - name: Lint with pylint
+      run: pylint --output-format="github" --enable="all" asyncmd
