### 2025-07-07 abb627e
Add linting to documentation
