+++ pyproject.toml
@@ -84,3 +84,10 @@ load-plugins = ["pylint.extensions.bad_builtin",
 # Minimum Python version to use for version dependent checks. Will default to the
 # version used to run pylint.
 py-version = "3.10"
+
+[tool.pylint."messages control"]
+# Enable the message, report, category or checker with the given id(s). You can
+# either give multiple identifier separated by comma (,) or put this option
+# multiple time (only on the command line, not in the configuration file where it
+# should appear only once). See also the "--disable" option for examples.
+enable = ["all"]
