### 2025-07-07 2682326
WIP: refactor to remove pylint errors
