+++ src/asyncmd/trajectory/convert.py
@@ -12,9 +12,14 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with asyncmd. If not, see <https://www.gnu.org/licenses/>.
+"""
+This module contains various classes used for trajectory extraction and concatenation.
+
+Most notably the FrameExtractor classes and the TrajectoryConcatenator.
+"""
 import os
 import abc
-import typing
+import collections.abc
 import asyncio
 import logging
 import functools
@@ -48,8 +53,8 @@ def _is_documented_by(original):
 
 def _attach_mda_trafos_to_universe(
         universe: mda.Universe,
-        mda_transformations: typing.Optional[list[typing.Callable]] = None,
-        mda_transformations_setup_func: typing.Optional[typing.Callable] = None,
+        mda_transformations: list[collections.abc.Callable] | None = None,
+        mda_transformations_setup_func: collections.abc.Callable | None = None,
                                   ) -> mda.Universe:
     """
     Attach MDAnalysis transformations to a given universe.
@@ -64,10 +69,10 @@ def _attach_mda_trafos_to_universe(
     ----------
     universe : MDAnalysis.core.universe.Universe
         The universe to attach the transformations to.
-    mda_transformations : typing.Optional[list[typing.Callable]], optional
+    mda_transformations : list[collections.abc.Callable] or None, optional
         List of MDAnalysis transformations to attach, by default None
-    mda_transformations_setup_func : typing.Optional[typing.Callable], optional
-        Setup function to attach MDAnalysis transformatiosn to the universe,
+    mda_transformations_setup_func : collections.abc.Callable or None, optional
+        Setup function to attach MDAnalysis transformations to the universe,
         by default None
 
     Returns
@@ -78,7 +83,7 @@ def _attach_mda_trafos_to_universe(
     Raises
     ------
     ValueError
-        If both ``mda_transformations`` and ``mda_transformations_setupt_func``
+        If both ``mda_transformations`` and ``mda_transformations_setup_func``
         are given.
     """
     # NOTE: this func is used to attach the MDAnalysis transformations to
@@ -120,8 +125,8 @@ class TrajectoryConcatenator:
 
     def __init__(self,
         invert_v_for_negative_step: bool = True,
-        mda_transformations: typing.Optional[list[typing.Callable]] = None,
-        mda_transformations_setup_func: typing.Optional[typing.Callable] = None,
+        mda_transformations: list[collections.abc.Callable] | None = None,
+        mda_transformations_setup_func: collections.abc.Callable | None = None,
                  ) -> None:
         """
         Initialize a :class:`TrajectoryConcatenator`.
@@ -163,7 +168,7 @@ class TrajectoryConcatenator:
         self.mda_transformations_setup_func = mda_transformations_setup_func
 
     def concatenate(self, trajs: "list[Trajectory]", slices: "list[tuple]",
-                    tra_out: str, struct_out: typing.Optional[str] = None,
+                    tra_out: str, struct_out: str | None = None,
                     overwrite: bool = False,
                     remove_double_frames: bool = True) -> Trajectory:
         """
@@ -186,7 +191,7 @@ class TrajectoryConcatenator:
             Whether we should overwrite existing output trajectories,
             by default False.
         remove_double_frames : bool, optional
-            Wheter we should try to remove double frames from the concatenated
+            Whether we should try to remove double frames from the concatenated
             output trajectory.
             Note that we use a simple heuristic to determine double frames,
             we just check if the integration time is the same for both frames,
@@ -224,11 +229,15 @@ class TrajectoryConcatenator:
             mda_transformations_setup_func=self.mda_transformations_setup_func,
             )
         start0, stop0, step0 = slices[0]
-        if remove_double_frames:
-            last_time_seen = None
[…]
