+++ .github/workflows/lint-with-pylint.yml
@@ -24,4 +24,4 @@ jobs:
        python -m pip install -e .
        python -m pip install pylint
     - name: Lint with pylint
-      run: pylint --output-format="github" --enable="all" asyncmd
+      run: pylint --output-format="github" asyncmd
