### 2025-07-07 5417465
WIP: refactoring for pylint
