+++ src/asyncmd/trajectory/propagate.py
@@ -16,14 +16,15 @@ import os
 import asyncio
 import aiofiles
 import aiofiles.os
+import collections.abc
 import inspect
 import logging
-import typing
 import numpy as np
 
 from .trajectory import Trajectory
 from .functionwrapper import TrajectoryFunctionWrapper
 from .convert import TrajectoryConcatenator
+from ..mdengine import MDEngine
 from ..utils import (get_all_traj_parts,
                      get_all_file_parts,
                      nstout_from_mdconfig,
@@ -53,11 +54,11 @@ async def construct_TP_from_plus_and_minus_traj_segments(
                                 plus_state: int,
                                 state_funcs: "list[TrajectoryFunctionWrapper]",
                                 tra_out: str,
-                                struct_out: typing.Optional[str] = None,
+                                struct_out: str | None = None,
                                 overwrite: bool = False
                                                          ) -> Trajectory:
     """
-    Construct a continous TP from plus and minus segments until states.
+    Construct a continuous TP from plus and minus segments until states.
 
     This is used e.g. in TwoWay TPS or if you try to get TPs out of a committor
     simulation. Note, that this inverts all velocities on the minus segments.
@@ -179,6 +180,13 @@ class _TrajectoryPropagator:
     # (private) superclass for InPartsTrajectoryPropagator and
     # ConditionalTrajectoryPropagator,
     # here we keep the common functions shared between them
+    def __init__(self, engine_cls: type[MDEngine], engine_kwargs: dict,
+                 walltime_per_part: float,
+                 ) -> None:
+        self.engine_cls = engine_cls
+        self.engine_kwargs = engine_kwargs
+        self.walltime_per_part = walltime_per_part
+
     async def remove_parts(self, workdir: str, deffnm: str,
                            file_endings_to_remove: list[str] = ["trajectories",
                                                                 "log"],
@@ -189,7 +197,7 @@ class _TrajectoryPropagator:
         Remove all `$deffnm.part$num.$file_ending` files for file_endings.
 
         Can be useful to clean the `workdir` from temporary files if e.g. only
-        the concatenate trajectory is of interesst (like in TPS).
+        the concatenate trajectory is of interest (like in TPS).
 
         Parameters
         ----------
@@ -287,7 +295,7 @@ class InPartsTrajectoryPropagator(_TrajectoryPropagator):
     Useful to make full use of backfilling with short(ish) simulation jobs and
     also to run simulations that are longer than the timelimit.
     """
-    def __init__(self, n_steps: int, engine_cls,
+    def __init__(self, n_steps: int, engine_cls: type[MDEngine],
                  engine_kwargs: dict, walltime_per_part: float,
                  ) -> None:
         """
@@ -304,10 +312,9 @@ class InPartsTrajectoryPropagator(_TrajectoryPropagator):
         walltime_per_part : float
             Walltime per trajectory segment, in hours.
         """
+        super().__init__(engine_cls=engine_cls, engine_kwargs=engine_kwargs,
+                         walltime_per_part=walltime_per_part)
         self.n_steps = n_steps
-        self.engine_cls = engine_cls
-        self.engine_kwargs = engine_kwargs
-        self.walltime_per_part = walltime_per_part
 
     async def propagate_and_concatenate(self,
                                         starting_configuration: Trajectory,
@@ -365,7 +372,7 @@ class InPartsTrajectoryPropagator(_TrajectoryPropagator):
         """
         Propagate the trajectory until self.n_steps integration are done.
 
-        Return a list of trajecory segments and the first condition fullfilled.
+        Return a list of trajectory segments and the first condition fulfilled.
 
         Parameters
         ----------
[…]
