+++ docs/source/community/contributing.md
@@ -52,7 +52,7 @@ asyncmd uses [pylint] to perform linting and ensure code quality. It will also b
 ```{important}
 The current configuration fails only if [pylint] finds an [error](https://pylint.readthedocs.io/en/latest/user_guide/messages/messages_overview.html#error-category) or if the [pylint] rating decreases to below 8.2.
 Note that this somewhat relaxed setting is due to historical code and will be tightened in the future as soon as we refactored the old code.
-All new code should strive to not add any [pylint] messages, but at least not result in additional error or warning messages.
+**All new code should strive to not add any [pylint] messages, but at least not result in additional error or warning messages.**
 ```
 
 You can run pytest using something like the following:
