### 2025-07-08 eec87bf
Add module/file-level docstrings for everything; tighten pylint configuration a bit
