### 2025-07-08 2401469
more tests for ConditionalTrajectoryPropagator
