### 2025-07-08 4bb6088
Add missing file-level docstring
