### 2025-07-08 add6572
Docs: spelling fixes
