+++ docs/source/guide/frame_extraction_and_trajectory_concatenation.md
@@ -54,7 +54,7 @@ It also contains a basic example on how to write your own {py:class}`FrameExtrac
 The {py:mod}`asyncmd.trajectory.convert` module furthermore contains a class to concatenate {py:class}`Trajectory <asyncmd.Trajectory>` segments, the {py:class}`TrajectoryConcatenator <asyncmd.trajectory.TrajectoryConcatenator>`.
 Again, it is reexported in the {py:mod}`asyncmd.trajectory` module for convenient imports.
 It can be used to concatenate lists of trajectory segments in any order (and possibly with inverted momenta) by passing a list of trajectory segments and a list of tuples (slices) that specify the frames to use in the concatenated output trajectory.
-Note that this class gives you all customizability at the cost of complexity, if you just want to construct a transition from trajectory segments, the {py:func}`<construct_TP_from_plus_and_minus_traj_segments <asyncmd.trajectory.construct_TP_from_plus_and_minus_traj_segments>` function is most probably easier to use and what you want (it uses the {py:class}`TrajectoryConcatenator <asyncmd.trajectory.convert.TrajectoryConcatenator>` under the hood anyway).
+Note that this class gives you all customizability at the cost of complexity, if you just want to construct a transition from trajectory segments, the {py:func}`construct_TP_from_plus_and_minus_traj_segments <asyncmd.trajectory.construct_TP_from_plus_and_minus_traj_segments>` function is most probably easier to use and what you want (it uses the {py:class}`TrajectoryConcatenator <asyncmd.trajectory.convert.TrajectoryConcatenator>` under the hood anyway).
 
 ```{note}
 The {py:class}`TrajectoryConcatenator <asyncmd.trajectory.TrajectoryConcatenator>` also has an async version of the function doing the work ({py:meth}`TrajectoryConcatenator.concatenate <asyncmd.trajectory.TrajectoryConcatenator.concatenate>` vs {py:meth}`TrajectoryConcatenator.concatenate <asyncmd.trajectory.TrajectoryConcatenator.concatenate_async>` respectively).
