### 2025-07-10 f2e8e5c
Add 'state of the field' section to joss publication
