+++ joss_paper/paper.bib
@@ -171,3 +171,116 @@ year = {2011}
   journal = {GitHub repository},
   url = {https://github.com/bio-phys/aimmd}
 }
+@article{ops1,
+  author = {Swenson, David W. H. and Prinz, Jan-Hendrik and Noe, Frank and Chodera, John D. and Bolhuis, Peter G.},
+  title = {{OpenPathSampling}: {A} {Python} Framework for Path Sampling Simulations. 1. {Basics}},
+  journal = {Journal of Chemical Theory and Computation},
+  volume = {15},
+  number = {2},
+  pages = {813-836},
+  year = {2019},
+  doi = {10.1021/acs.jctc.8b00626},
+  note ={PMID: 30336030},
+  URL = {https://doi.org/10.1021/acs.jctc.8b00626},
+  eprint = {https://doi.org/10.1021/acs.jctc.8b00626},
+}
+@article{ops2,
+  author = {Swenson, David W. H. and Prinz, Jan-Hendrik and Noe, Frank and Chodera, John D. and Bolhuis, Peter G.},
+  title = {{OpenPathSampling}: {A} {P}ython Framework for Path Sampling Simulations. 2. {B}uilding and Customizing Path Ensembles and Sample Schemes},
+  journal = {Journal of Chemical Theory and Computation},
+  volume = {15},
+  number = {2},
+  pages = {837-856},
+  year = {2019},
+  doi = {10.1021/acs.jctc.8b00627},
+  URL = {https://doi.org/10.1021/acs.jctc.8b00627},
+  eprint = {https://doi.org/10.1021/acs.jctc.8b00627},
+}
+@article{irrgang2018gmxapi,
+  author = {Irrgang, M Eric and Hays, Jennifer M and Kasson, Peter M},
+  title = {gmxapi: a high-level interface for advanced control and extension of molecular dynamics simulations},
+  journal = {Bioinformatics},
+  volume = {34},
+  number = {22},
+  pages = {3945-3947},
+  year = {2018},
+  month = {06},
+  issn = {1367-4803},
+  doi = {10.1093/bioinformatics/bty484},
+  url = {https://doi.org/10.1093/bioinformatics/bty484},
+  eprint = {https://academic.oup.com/bioinformatics/article-pdf/34/22/3945/48921266/bioinformatics\_34\_22\_3945.pdf},
+}
+@article{irrgang2022gmxapi,
+  doi = {10.1371/journal.pcbi.1009835},
+  author = {Irrgang, M. Eric AND Davis, Caroline AND Kasson, Peter M.},
+  journal = {PLOS Computational Biology},
+  publisher = {Public Library of Science},
+  title = {gmxapi: A GROMACS-native Python interface for molecular dynamics with ensemble and plugin support},
+  year = {2022},
+  month = {02},
+  volume = {18},
+  url = {https://doi.org/10.1371/journal.pcbi.1009835},
+  pages = {1-12},
+  number = {2},
+}
+@article{huber2020aiida,
+  title = {{{AiiDA}} 1.0, a Scalable Computational Infrastructure for Automated Reproducible Workflows and Data Provenance},
[…]
