### 2025-07-14 a23ec67
update docs to make clear we dont use LFS anymore
