### 2025-07-14 5600d48
Reduce number of pylint messages for configuration module
