### 2025-07-14 785bed6
config, tools, and utils: Reduce number of pylint messages
