+++ src/asyncmd/tools.py
@@ -66,7 +66,7 @@ def ensure_executable_available(executable: str) -> str:
         # see if we find it in $PATH
         executable = shutil.which(executable)
     else:
-        raise ValueError(f"{executable} must be an existing path or accesible "
+        raise ValueError(f"{executable} must be an existing path or accessible "
                          + "via the $PATH environment variable.")
     return executable
 
@@ -83,7 +83,6 @@ def remove_file_if_exist(f: str):
     try:
         os.remove(f)
     except FileNotFoundError:
-        # TODO: should we info/warn if the file is not there?
         pass
 
 
@@ -99,14 +98,13 @@ async def remove_file_if_exist_async(f: str):
     try:
         await aiofiles.os.remove(f)
     except FileNotFoundError:
-        # TODO: should we info/warn if the file is not there?
         pass
 
 
 class FlagChangeList(collections.abc.MutableSequence):
     """A list that knows if it has been changed after initializing."""
 
-    def __init__(self, data: typing.Iterable) -> None:
+    def __init__(self, data: collections.abc.Iterable) -> None:
         """
         Initialize a `FlagChangeList`.
 
@@ -165,10 +163,10 @@ class FlagChangeList(collections.abc.MutableSequence):
         self._data.insert(index, value)
         self._changed = True
 
-    def __add__(self, other: typing.Iterable):
+    def __add__(self, other: collections.abc.Iterable):
         return FlagChangeList(data=self._data + list(other))
 
-    def __iadd__(self, other: typing.Iterable):
+    def __iadd__(self, other: collections.abc.Iterable):
         for val in other:
             self.append(val)
         return self
@@ -183,7 +181,7 @@ class TypedFlagChangeList(FlagChangeList):
     `data=["abc"]` (and not `data=["a", "b", "c"]`).
     """
 
-    def __init__(self, data: typing.Iterable, dtype: typing.Type) -> None:
+    def __init__(self, data: collections.abc.Iterable, dtype: type) -> None:
         """
         Initialize a `TypedFlagChangeList`.
 
@@ -193,7 +191,7 @@ class TypedFlagChangeList(FlagChangeList):
             (Initial) data for this `TypedFlagChangeList`.
         dtype : Callable datatype
             The datatype for all entries in this `TypedFlagChangeList`. Will be
-            called on every value seperately and is expected to convert to the
+            called on every value separately and is expected to convert to the
             desired datatype.
         """
         self._dtype = dtype  # set first to use in _convert_type method
@@ -203,19 +201,19 @@ class TypedFlagChangeList(FlagChangeList):
         super().__init__(data=typed_data)
 
     @property
-    def dtype(self) -> typing.Type:
+    def dtype(self) -> type:
         """
         All values in this `TypedFlagChangeList` are converted to dtype.
 
         Returns
         -------
-        typing.Type
+        type
         """
         return self._dtype
 
-    def _ensure_iterable(self, data) -> typing.Iterable:
+    def _ensure_iterable(self, data) -> collections.abc.Iterable:
         if getattr(data, '__iter__', None) is None:
-            # convienience for singular options,
+            # convenience for singular options,
             # if it has no iter attribute we assume it is the only item
             data = [data]
         elif isinstance(data, str):
@@ -225,7 +223,7 @@ class TypedFlagChangeList(FlagChangeList):
         return data
 
     def _convert_type(self, value,
-                      index: typing.Optional[int | slice] = None) -> list:
+                      index: int | slice | None = None) -> list:
         # here we ignore index, but passing it should in principal make it
         # possible to use different dtypes for different indices
         if isinstance(index, int):
@@ -252,14 +250,14 @@ class TypedFlagChangeList(FlagChangeList):
[…]
