### 2025-07-14 a3b1e55
mdconfig: refactor to reduce number of pylint messages
