### 2025-07-14 d6991e5
mdengine: make pylint happy and improve docstrings
