+++ src/asyncmd/mdengine.py
@@ -23,12 +23,10 @@ from .trajectory.trajectory import Trajectory
 
 class EngineError(Exception):
     """Exception raised when something goes wrong with the (MD)-Engine."""
-    pass
 
 
 class EngineCrashedError(EngineError):
     """Exception raised when the (MD)-Engine crashes during a run."""
-    pass
 
 
 class MDEngine(abc.ABC):
@@ -38,64 +36,127 @@ class MDEngine(abc.ABC):
     @abc.abstractmethod
     async def apply_constraints(self, conf_in: Trajectory,
                                 conf_out_name: str) -> Trajectory:
-        # apply constraints to given conf_in, write conf_out_name and return it
+        """
+        Apply constraints to given conf_in, write conf_out_name and return it.
+
+        Parameters
+        ----------
+        conf_in : Trajectory
+        conf_out_name : str
+
+        Returns
+        -------
+        Trajectory
+        """
         raise NotImplementedError
 
     @abc.abstractmethod
-    # TODO: think about the most general interface!
-    # NOTE: We assume that we do not change the system for/in one engine,
-    #       i.e. .top, .ndx, mdp-object, ...?! should go into __init__
     async def prepare(self, starting_configuration: Trajectory, workdir: str,
                       deffnm: str) -> None:
+        """
+        Prepare the engine to run a MD from starting_configuration.
+
+        NOTE: We assume that we do not change the system for/in one engine,
+        i.e. .top, .ndx, mdp-object, ...?! should go into __init__.
+
+        Parameters
+        ----------
+        starting_configuration : Trajectory
+            The initial configuration.
+        workdir : str
+            The directory in which the MD will be performed.
+        deffnm : str
+            The standard filename to use for this MD run.
+        """
         raise NotImplementedError
 
     @abc.abstractmethod
-    # TODO: should this be a classmethod?
     #@classmethod
     async def prepare_from_files(self, workdir: str, deffnm: str) -> None:
-        # this should prepare the engine to continue a previously stopped simulation
-        # starting with the last trajectory part in workdir that is compatible with deffnm
+        """
+        Prepare the engine to continue a previously stopped simulation starting
+        with the last trajectory part in workdir that is compatible with deffnm.
+
+        NOTE: This can not be a classmethod (reliably) because we set top/ndx/
+        mdconfig/etc in '__init__'.
+
+        Parameters
+        ----------
+        workdir : str
+            The directory in which the MD will be/ was previously performed.
+        deffnm : str
+            The standard filename to use for this MD run.
+        """
         raise NotImplementedError
 
     @abc.abstractmethod
     async def run_walltime(self, walltime: float) -> Trajectory:
-        # run for specified walltime
-        # NOTE: must be possible to run this multiple times after preparing once!
+        """
+        Run for specified walltime.
+
+        NOTE: Must be possible to run this multiple times after preparing once!
+
+        Parameters
+        ----------
+        walltime : float
+
+        Returns
+        -------
+        Trajectory
+        """
         raise NotImplementedError
 
     @abc.abstractmethod
     async def run_steps(self, nsteps: int,
                         steps_per_part: bool = False) -> Trajectory:
-        # run for specified number of steps
-        # NOTE: not sure if we need it, but could be useful
-        # NOTE: make sure we can run multiple times after preparing once!
+        """
+        Run for a specified number of integration steps.
+
+        NOTE: Make sure we can run multiple times after preparing once!
+
+        Parameters
+        ----------
+        nsteps : int
+        steps_per_part : bool, optional
+            Count nsteps for this part/run or in total, by default False
+
+        Returns
+        -------
+        Trajectory
+        """
         raise NotImplementedError
 
     @property
     @abc.abstractmethod
     def current_trajectory(self) -> Trajectory | None:
-        # return current trajectory: Trajectory or None
[…]
