### 2025-07-14 10136e6
gromacs.mdconfig and gromacs.utils: make pylint happy
