### 2025-07-15 5bc459c
trajectory.convert: refactor to increase usability, readbility and make pylint happy
