+++ src/asyncmd/trajectory/convert.py
@@ -110,7 +110,9 @@ class TrajectoryConcatenator:
     and returns one trajectory containing only the selected frames in the order
     specified by the slices.
     Velocities are automatically inverted if the step of a slice is negative,
-    this can be controlled via the invert_v_for_negative_step attribute.
+    this can be controlled via the ``invert_v_for_negative_step`` attribute.
+    Double frames are also automatically removed, which can be controlled via
+    the ``remove_double_frames`` attribute.
     We assume that all trajs have the same structure file and attach the
     structure of the first traj if not told otherwise.
     Note that you can pass MDAnalysis transformations to this class to
@@ -121,12 +123,20 @@ class TrajectoryConcatenator:
     ----------
     invert_v_for_negative_step : bool
         Whether to invert all momenta for segments with negative stride.
+    remove_double_frames : bool
+        Whether we should (try to) remove double frames from the concatenated
+        output trajectory.
+        Note that a simple heuristic is used to determine double frames,
+        frames count as double if the integration time is the same for both
+        frames.
     """
 
-    def __init__(self,
-        invert_v_for_negative_step: bool = True,
-        mda_transformations: list[collections.abc.Callable] | None = None,
-        mda_transformations_setup_func: collections.abc.Callable | None = None,
+    def __init__(
+            self,
+            invert_v_for_negative_step: bool = True,
+            remove_double_frames: bool = True, *,
+            mda_transformations: list[collections.abc.Callable] | None = None,
+            mda_transformations_setup_func: collections.abc.Callable | None = None,
                  ) -> None:
         """
         Initialize a :class:`TrajectoryConcatenator`.
@@ -136,6 +146,9 @@ class TrajectoryConcatenator:
         invert_v_for_negative_step : bool, optional
             Whether to invert all momenta for segments with negative stride,
             by default True.
+        remove_double_frames : bool, optional
+            Whether we should (try to) remove double frames from the concatenated
+            output trajectory, by default True.
         mda_transformations : list of callables, optional
             If given will be added as a list of transformations to the
             MDAnalysis universe as
@@ -143,7 +156,8 @@ class TrajectoryConcatenator:
             See the ``mda_transformations_setup_func`` argument if your
             transformations need additional universe-dependant arguments, e.g.
             atomgroups from the universe.
-            See https://docs.mdanalysis.org/stable/documentation_pages/trajectory_transformations.html
+            See
+            https://docs.mdanalysis.org/stable/documentation_pages/trajectory_transformations.html
             for more on MDAnalysis transformations.
         mda_transformations_setup_func: callable, optional
             If given will be called to attach user-defined MDAnalysis
@@ -154,12 +168,16 @@ class TrajectoryConcatenator:
             after defining ``list_of_trafos`` (potentially depending on the
             universe or atomgroups therein) and then finally returns the
             universe with trafos.
-            See https://docs.mdanalysis.org/stable/documentation_pages/trajectory_transformations.html
+            See
+            https://docs.mdanalysis.org/stable/documentation_pages/trajectory_transformations.html
             for more on MDAnalysis transformations.
         """
         self.invert_v_for_negative_step = invert_v_for_negative_step
-        if (mda_transformations is not None
-            and mda_transformations_setup_func is not None):
+        self.remove_double_frames = remove_double_frames
+        if (
+            mda_transformations is not None
+            and mda_transformations_setup_func is not None
+        ):
             raise ValueError("`mda_transformations` and "
                              "`mda_transformations_setup_func` are mutually "
                              "exclusive, but both were given."
[…]
