+++ pyproject.toml
@@ -92,3 +92,9 @@ py-version = "3.10"
 # multiple time (only on the command line, not in the configuration file where it
 # should appear only once). See also the "--disable" option for examples.
 enable = ["all"]
+
+[tool.pylint.design]
+# increase maximum number of arguments for functions/methods (default=5)
+max-args = 6
+# but decrease maximum number of positional arguments (default=5)
+max-positional-arguments = 4
