### 2025-07-15 995e0b1
Add some tests for utils dispatching to the correct engine submodule
