### 2025-07-20 538f0c0
Big overhaul of Trajectory class (mostly under the hood), completely rework caching of wrapped function values; also refactor to remove doublicate code used for class initialization all over asyncmd to make pylint happy
