+++ src/asyncmd/trajectory/trajectory.py
@@ -17,21 +17,31 @@ This module contains the implementation the asyncmd.Trajectory class.
 
 It also contains some helper function related to the global Trajectory registry
 used for trajectory function value caching.
+The actual :class:`TrajectoryFunctionValueCache` classes can be found in the
+``trajectory_cache`` module.
 """
-import io
-import os
-import copy
-import typing
 import asyncio
-import hashlib
-import logging
-import zipfile
 import collections
-import numpy as np
-import MDAnalysis as mda
+import dataclasses
+import hashlib
+import io
+import logging
+import os
+import typing
 
+import MDAnalysis as mda
+import numpy as np
 
 from .._config import _GLOBALS
+from .trajectory_cache import (TrajectoryFunctionValueCache,
+                               TrajectoryFunctionValueCacheInH5PY,
+                               TrajectoryFunctionValueCacheInMemory,
+                               TrajectoryFunctionValueCacheInNPZ,
+                               ValuesAlreadyStoredError)
+
+if typing.TYPE_CHECKING:  # pragma: no cover
+    # only import for typing to avoid circular imports
+    from .functionwrapper import TrajectoryFunctionWrapper
 
 
 logger = logging.getLogger(__name__)
@@ -40,20 +50,58 @@ logger = logging.getLogger(__name__)
 # dictionary in which we keep track of trajectory objects
 # we use it to always return the *same* object for the same trajectory (by hash)
 # this makes it easy to ensure that we never calculate CV functions twice
-_TRAJECTORIES_BY_HASH = {}
+_TRAJECTORIES_BY_HASH: dict[int, "Trajectory"] = {}
+
+
+def clear_all_cache_values_for_all_trajectories() -> None:
+    """
+    Clear all function values cached for each :class:`Trajectory` currently in existence.
+
+    For file-based caches, this also removes the associated cache files.
+    """
+    for traj in _TRAJECTORIES_BY_HASH.values():
+        traj.clear_all_cache_values()
+
+
+def _update_cache_type_for_all_trajectories(copy_content: bool = True,
+                                            clear_old_cache: bool = False,
+                                            ) -> None:
+    """
+    Update the cache type for each :class:`Trajectory` currently in existence.
+
+    By default the content of the current caches is copied to the new caches.
+    This will only have an effect if the globally set ``cache_type`` differs
+    from what each `Trajectory` currently uses.
+    See :func:`asyncmd.config.set_trajectory_cache_type` to set the ``cache_type``.
+    To clear the old/previously set caches (after copying their values), pass
+    ``clear_old_cache=True``.
+
+    Parameters
+    ----------
+    copy_content : bool, optional
+        Whether to copy the current cache content to the new cache,
+        by default True
+    clear_old_cache : bool, optional
+        Whether to clear the old/previously set cache, by default False.
+    """
+    for traj in _TRAJECTORIES_BY_HASH.values():
+        traj.update_cache_type(copy_content=copy_content,
+                               clear_old_cache=clear_old_cache,
+                               )
 
 
 def _forget_all_trajectories() -> None:
     """
     Forget about the existence of all :class:`Trajectory` objects.
 
-    This will result in new :class:`Trajectory` objects beeing created even for
-    the same underlying trajectory_files. Usualy you do not want this as it
-    results in unecessary calculations if the same wrapped and cached function
+    This will result in new :class:`Trajectory` objects being created even for
+    the same underlying trajectory_files. Usually you do not want this as it
+    results in unnecessary calculations if the same wrapped and cached function
     is applied to both objects. This function exists as a hidden function as it
     is used in the tests and it might be helpful under certain circumstances.
[…]
