+++ tests/trajectory/test_trajectory.py
@@ -17,15 +17,16 @@ import os
 import pickle
 import numpy as np
 
-from unittest.mock import AsyncMock
+from unittest.mock import Mock, PropertyMock
 
 import asyncmd
 from asyncmd import Trajectory
 from asyncmd.config import _GLOBALS  # noqa: F401
-from asyncmd.trajectory.trajectory import (TrajectoryFunctionValueCacheMEMORY,
-                                           TrajectoryFunctionValueCacheNPZ,
-                                           TrajectoryFunctionValueCacheH5PY,
-                                           )
+from asyncmd.trajectory.trajectory_cache import (TrajectoryFunctionValueCacheInMemory,
+                                                 TrajectoryFunctionValueCacheInNPZ,
+                                                 TrajectoryFunctionValueCacheInH5PY,
+                                                 ValuesAlreadyStoredError,
+                                                 )
 from asyncmd.trajectory.functionwrapper import TrajectoryFunctionWrapper
 
 
@@ -71,6 +72,23 @@ class Test_Trajectory(TBase):
     def setup_method(self):
         super().setup_method()
         asyncmd.trajectory._forget_all_trajectories()
+
+    @pytest.mark.parametrize(["trajectory_files", "structure_file"],
+                             [("tests/test_data/trajectory/ala_traj.trr",  # existing traj
+                               "tests/test_data/trajectory/NON_EXISTING_STRUCTURE",  # ...
+                               ),
+                              (["tests/test_data/trajectory/ala_traj.trr"],  # existing traj (list)
+                               "tests/test_data/trajectory/NON_EXISTING_STRUCTURE",  # ...
+                               ),
+                              # and the other way around: non-existing traj, but struct present
+                              ("tests/test_data/trajectory/NON_EXISTING_TRAJECTORY",
+                               "tests/test_data/trajectory/ala.tpr")
+                              ]
+                             )
+    def test_init_raises_non_existing_files(self, trajectory_files, structure_file):
+        with pytest.raises(FileNotFoundError):
+            _ = Trajectory(trajectory_files=trajectory_files,
+                           structure_file=structure_file)
 
     @pytest.mark.parametrize(["traj_file", "struct_file", "truth"],
                              [("tests/test_data/trajectory/ala_traj.trr",
@@ -296,7 +314,7 @@ class Test_Trajectory(TBase):
             return Trajectory(
                     trajectory_files="tests/test_data/trajectory/ala_traj.trr",
                     structure_file="tests/test_data/trajectory/ala.tpr",
-                              )
+                    )
 
         def assert_neq(traj1, traj2):
             # check both eq and neq at once
@@ -306,58 +324,42 @@ class Test_Trajectory(TBase):
         traj2 = make_traj()
         assert traj1 == traj2  # make sure they are equal to begin with
         assert not traj1 != traj2  # and check that neq also works
-        # modify hash
-        traj2 = make_traj()  # get a new traj2
-        traj2._traj_hash += 1
+        # get another traj2 to test [not] eq (for two trajs which are different)
+        traj2 = Trajectory(
+                    trajectory_files="tests/test_data/trajectory/ala_conf_in_no_state.trr",
+                    structure_file="tests/test_data/trajectory/ala.tpr",
+                    )
         assert_neq(traj1, traj2)
         # test for non trajectory objects
         assert_neq(traj1, object())
 
-    @pytest.mark.parametrize("default_cache_type",
-                             [None, "npz", "h5py", "memory"])
     @pytest.mark.parametrize("cache_type",
-                             [None, "npz", "h5py", "memory"])
+                             ["npz", "h5py", "memory"])
     @pytest.mark.parametrize("initial_cache_type",
[…]
