### 2025-07-21 91bf2c7
asyncmd.tracjectory.propagate: refactor to increase code quality and make pylint happier
