+++ src/asyncmd/trajectory/propagate.py
@@ -17,31 +17,29 @@ This module contains classes for propagation of MD in segments and/or until a co
 
 Most notable are the InPartsTrajectoryPropagator and the ConditionalTrajectoryPropagator.
 Also of interest especially in the context of path sampling could be the function
-construct_TP_from_plus_and_minus_traj_segments, which can be used directly on the
+construct_tp_from_plus_and_minus_traj_segments, which can be used directly on the
 output of a ConditionalTrajectoryPropagator to generate trajectories connecting two
 fulfilled conditions.
 """
-import os
 import asyncio
-import aiofiles
-import aiofiles.os
 import collections.abc
+import copy
 import inspect
 import logging
+import os
+import typing
+
+import aiofiles
+import aiofiles.os
 import numpy as np
 
-from .trajectory import Trajectory
-from .functionwrapper import TrajectoryFunctionWrapper
-from .convert import TrajectoryConcatenator
 from ..mdengine import MDEngine
-from ..utils import (get_all_traj_parts,
-                     get_all_file_parts,
-                     nstout_from_mdconfig,
-                     )
-from ..tools import (remove_file_if_exist,
-                     remove_file_if_exist_async,
-                     FlagChangeList,
-                     )
+from ..tools import FlagChangeList, remove_file_if_exist_async
+from ..utils import (get_all_file_parts, get_all_traj_parts,
+                     nstout_from_mdconfig)
+from .convert import TrajectoryConcatenator
+from .functionwrapper import TrajectoryFunctionWrapper
+from .trajectory import Trajectory
 
 
 logger = logging.getLogger(__name__)
@@ -52,20 +50,18 @@ class MaxStepsReachedError(Exception):
     Error raised when the simulation terminated because the (user-defined)
     maximum number of integration steps/trajectory frames has been reached.
     """
-    pass
 
 
-# TODO: move to trajectory.convert?
-async def construct_TP_from_plus_and_minus_traj_segments(
+async def construct_tp_from_plus_and_minus_traj_segments(
+                                *,
                                 minus_trajs: "list[Trajectory]",
                                 minus_state: int,
                                 plus_trajs: "list[Trajectory]",
                                 plus_state: int,
                                 state_funcs: "list[TrajectoryFunctionWrapper]",
                                 tra_out: str,
-                                struct_out: str | None = None,
-                                overwrite: bool = False
-                                                         ) -> Trajectory:
+                                **concatenate_kwargs,
+                                ) -> Trajectory:
     """
     Construct a continuous TP from plus and minus segments until states.
 
@@ -87,11 +83,15 @@ async def construct_TP_from_plus_and_minus_traj_segments(
         the minus and plus state indices!
     tra_out : str
         Absolute or relative path to the output trajectory file.
-    struct_out : str, optional
-        Absolute or relative path to the output structure file, if None we will
-        use the structure file of the first minus_traj, by default None.
-    overwrite : bool, optional
-        Whether we should overwrite tra_out if it exists, by default False.
+    concatenate_kwargs : dict
+        All (other) keyword arguments will be passed as is to the
+        :meth:`TrajectoryConcatenator.concatenate` method. These include, e.g.,
+
+        - struct_out : str, optional
+            Absolute or relative path to the output structure file, if None we will
+            use the structure file of the first minus_traj, by default None.
+        - overwrite : bool, optional
+            Whether we should overwrite tra_out if it exists, by default False.
 
     Returns
     -------
@@ -179,33 +179,34 @@ async def construct_TP_from_plus_and_minus_traj_segments(
                                                     trajs=trajs,
                                                     slices=slices,
                                                     tra_out=tra_out,
-                                                    struct_out=struct_out,
-                                                    overwrite=overwrite,
+                                                    **concatenate_kwargs
[…]
