+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Renamed `asyncmd.trajectory.construct_TP_from_plus_and_minus_traj_segments` to `asyncmd.trajectory.construct_tp_from_plus_and_minus_traj_segments`. Also minimal change to the call signature, arguments to the `TrajectoryConcatenator` class are now not explicit arguments anymore but instead all kwargs will be passed to the `TrajectoryConcatenator.concatenate` method.
+- `asyncmd.trajectory.propagate`: Changes to the call signatures of `TrajectoryPropagator` classes, the arguments that were previously only passed trough to the concatenator class are now not explicit anymore and instead all additional kwargs passed to the `TrajectoryPropagator.cut_and_concatenate` and `TrajectoryPropagator.propagate_and_concatenate` methods will be directly passed to the `concatenate` method of the `TrajectoryConcatenator`.
 - Renamed `asyncmd.config.set_default_trajectory_cache_type` to `asyncmd.config.set_trajectory_cache_type` and remove option to set cache type individually for `Trajectory` objects and instead always use the centrally configured cache. These changes were made while drastically simplifying the trajectory function value caching code (mostly under the hood) to increase maintainability and extensibility.
 - `TrajectoryFunctionWrapper` and their interaction with `Trajectories`: One notable semi-under-the-hood change is that `TrajectoryFunctionWrapper` subclasses now need to implement the `_get_values_for_trajectory` method instead of the `get_values_for_trajectory` method. The call signature of the function has not changed and the change should in all cases be as easy as appending the `_` to the function name in subclasses. This change was deemed necessary to avoid that users call the `get_values_for_trajectory` method expecting that the values would be cached (which they are not).
 - `TrajectoryConcatenator`: `remove_double_frames` is now an instance attribute instead of an argument to the `concatenate` and `concatenate_async` methods
