### 2025-07-21 b589b41
Slightly modify method signature of MDEngine ABC
