### 2025-07-22 0293317
GmxEgine(s): Refactor to make more readable and pylinat happier, improve  method for trr trajectories
