### 2025-07-22 dbdfb7e
tighten pylint configuration
