+++ pyproject.toml
@@ -61,7 +61,7 @@ where = ["src"]
 fail-on = ["E"]
 
 # Specify a score threshold under which the program will exit with error.
-fail-under = 8.6
+fail-under = 9.6
 
 # List of plugins (as comma separated values of python module names) to load,
 # usually to register additional checkers.
