### 2025-07-22 3a00ff4
Release v0.4.0
