+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2025-07-22
+
 ### Added
 
 - Pylint configuration (in `pyproject.toml`), automatically run pylint as a github action
@@ -49,6 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release on PyPi
 
-[unreleased]: https://github.com/bio-phys/asyncmd/compare/v0.3.3...HEAD
+[unreleased]: https://github.com/bio-phys/asyncmd/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/bio-phys/asyncmd/compare/v0.3.3...v0.4.0
 [0.3.3]: https://github.com/bio-phys/asyncmd/compare/v0.3.2...v0.3.3
 [0.3.2]: https://github.com/bio-phys/asyncmd/releases/tag/v0.3.2
