### 2025-07-22 #8321 @espottesmith
Paper comments:
- It doesn't seem to me (perhaps I'm missing something) that your code is specific to biochemistry,
- Your "Statement of Need" jumps right into what your code does without pausing to explain why you needed to write this code (this is covered somewhat in the summary, but it should also be present in "Statement of Need")
- Re "State of the Field", you might want to look at Atomate2, which has support for high-throughput [OpenMM](https://github.com/materialsproject/atomate2/tree/main/src/atomate2/openmm) workflows. I suspect that asyncmd has some unique features not present in Atomate2 (I'm not an Atomate2 user or developer, nor would I consider myself an MD expert, so I can't say for certain), but it seems to be trying to fill a similar niche to asyncmd and/or Aiida.
- It seems that one of your selling points is modularity; _i.e._, you provide a low-level interface which can then be used in arbitrary enhanced sampling methods. It would be nice (though perhaps not _necessary_) if there was an example where you use asyncmd to such an effect.
- Please fix the reference issues noted above.
- I don't see any glaring spelling or grammar issues, though I'll do a final check before acceptance
