### 2025-07-22 1970bfe
attempt to fix citations
