+++ joss_paper/paper.bib
@@ -247,7 +247,7 @@ year = {2011}
   pages = {110086},
   year = {2021},
   issn = {0927-0256},
-  doi = {https://doi.org/10.1016/j.commatsci.2020.110086},
+  doi = {10.1016/j.commatsci.2020.110086},
   url = {https://www.sciencedirect.com/science/article/pii/S0927025620305772},
   author = {Martin Uhrin and Sebastiaan P. Huber and Jusong Yu and Nicola Marzari and Giovanni Pizzi},
   keywords = {Data management, Database, Data sharing, Provenance, Computational workflows, Event-based, Robust computation, High-throughput},
@@ -271,10 +271,14 @@ year = {2011}
   url = {https://github.com/glotzerlab/row}
 }
 @article{bereau2024martignac,
-  title={Martignac: Computational workflows for reproducible, traceable, and composable coarse-grained Martini simulations},
-  author={Bereau, Tristan and Walter, Luis J and Rudzinski, Joseph F},
-  journal={arXiv preprint arXiv:2409.15478},
-  year={2024}
+  author = {Bereau, Tristan and Walter, Luis J. and Rudzinski, Joseph F.},
+  title = {Martignac: Computational Workflows for Reproducible, Traceable, and Composable Coarse-Grained Martini Simulations},
+  journal = {Journal of Chemical Information and Modeling},
+  volume = {64},
+  number = {24},
+  pages = {9413-9423},
+  year = {2024},
+  doi = {10.1021/acs.jcim.4c01754},
 }
 @misc{aiida-gromacs,
   author = {James Gebbie-Rayet and Jas Kalayan},
