### 2025-07-22 0266bf3
Add py.typed file to support type checking
