+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- "py.typed" file to support type checking
+
 ## [0.4.0] - 2025-07-22
 
 ### Added
