### 2025-07-23 99f31bd
WIP: refactor slurm.py
