### 2025-07-23 0efc118
Restored duplicated slurm.py
