+++ src/asyncmd/slurm.py
@@ -1,0 +1,1333 @@
+# This file is part of asyncmd.
+#
+# asyncmd is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# asyncmd is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with asyncmd. If not, see <https://www.gnu.org/licenses/>.
+"""
+This module contains the implementation of the classes to interact with Slurm.
+
+The SlurmClusterMediator is a singleton class (handling all sacct calls in a
+coordinated fashion) for all SlurmProcess instances.
+The SlurmProcess is a drop-in replacement for asyncio.subprocess.Subprocess and
+in this spirit this module also contains the function create_slurmprocess_submit,
+which similarly to asyncio.create_subprocess_exec, creates a SlurmProcess and
+directly submits the job.
+Finally this module contains two functions to set the configuration of this module,
+set_all_slurm_settings and set_slurm_settings.
+"""
+import asyncio
+import collections
+import logging
+import re
+import shlex
+import subprocess
+import time
+import typing
+import os
+import aiofiles
+import aiofiles.os
+
+from .tools import (ensure_executable_available,
+                    remove_file_if_exist_async,
+                    remove_file_if_exist,
+                    )
+from .tools import attach_kwargs_to_object as _attach_kwargs_to_object
+from ._config import _SEMAPHORES
+
+
+logger = logging.getLogger(__name__)
+
+
+class SlurmError(RuntimeError):
+    """Generic error superclass for all SLURM errors."""
+
+
+class SlurmCancellationError(SlurmError):
+    """Error raised when something goes wrong canceling a SLURM job."""
+
+
+class SlurmSubmissionError(SlurmError):
+    """Error raised when something goes wrong submitting a SLURM job."""
+
+
+# rudimentary map for slurm state codes to int return codes for poll
+# NOTE: these are the sacct states (they differ from the squeue states)
+#       cf. https://slurm.schedmd.com/sacct.html#lbAG
+#       and https://slurm.schedmd.com/squeue.html#lbAG
+# NOTE on error codes:
+#      we return:
+#       - None if the job has not finished
+#       - 0 if it completed successfully
+#       - 1 if the job failed (probably) due to user error (or we dont know)
+#       - 2 if the job failed (almost certainly) due to cluster/node-issues as
+#         recognized/detected by slurm
+_SLURM_STATE_TO_EXITCODE = {
+    "BOOT_FAIL": 1,  # Job terminated due to launch failure
+    # Job was explicitly cancelled by the user or system administrator.
+    "CANCELLED": 1,
+    # Job has terminated all processes on all nodes with an exit code of
+    # zero.
+    "COMPLETED": 0,
+    "DEADLINE": 1,  # Job terminated on deadline.
+    # Job terminated with non-zero exit code or other failure condition.
+    "FAILED": 1,
+    # Job terminated due to failure of one or more allocated nodes.
+    "NODE_FAIL": 2,
+    "OUT_OF_MEMORY": 1,  # Job experienced out of memory error.
+    "PENDING": None,  # Job is awaiting resource allocation.
+    # NOTE: preemption means interrupting a process to later restart it,
+    #       i.e. None is probably the right thing to return
+    "PREEMPTED": None,  # Job terminated due to preemption.
[…]
