### 2025-07-24 #8321 @espottesmith
Re: the biochem/biomolecular angle, what I meant (I forgot to finish that point) was that, in introducing your code, you don't need to focus so heavily on biochem. Your code is more general; you should advertise that!

Atomate2 is definitely more quantum chemical. I mentioned it only because it allows for high-throughput MD. If you don't think it's worth mentioning, I think that's fair.

I did miss that example. Thank you for pointing it out. I think, if you could make a small reference to what examples exist/what kinds of simulations your code has been used for, that would be helpful.
