### 2025-07-25 751a47c
WIP: refactor slurm module to make pylint happier and the code more readable
