+++ src/asyncmd/slurm/__init__.py
@@ -1,0 +1,19 @@
+# This file is part of asyncmd.
+#
+# asyncmd is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# asyncmd is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with asyncmd. If not, see <https://www.gnu.org/licenses/>.
+"""
+This module contains classes and functions for interacting with the SLURM queuing system.
+"""
+from .process import SlurmProcess, create_slurmprocess_submit
+from .config import set_all_slurm_settings, set_slurm_settings
