### 2025-07-27 f1c0ccf
Add some tests for SlurmProcess.submit
