### 2025-07-27 adb4ee7
Refactor slurm module to make pylint a bit happier and configuration behavior more consistent
