+++ src/asyncmd/slurm/process.py
@@ -29,7 +29,10 @@ import os
 import aiofiles
 
 from .cluster_mediator import SlurmClusterMediator
-from .constants_and_errors import SlurmCancellationError, SlurmSubmissionError
+from .constants_and_errors import (SlurmCancellationError,
+                                   SlurmSubmissionError,
+                                   SlurmError,
+                                   )
 from ..tools import (ensure_executable_available,
                      remove_file_if_exist_async,
                      remove_file_if_exist,
@@ -55,18 +58,29 @@ class SlurmProcess:
         Name or path to the scancel executable, by default "scancel".
     sleep_time : int
         Time (in seconds) between checks if the underlying job has finished
-        when using `self.wait`.
+        when using `self.wait`. By default 15 s.
+    stdfiles_removal : str
+        Whether to remove the stdout, stderr (and possibly stdin) files.
+        Possible values are:
+
+        - "success": remove on successful completion, i.e. zero returncode)
+        - "no": never remove
+        - "yes"/"always": remove on job completion independent of
+          returncode and also when using :meth:`terminate`
+
+        By default "success".
     """
 
     # use same instance of class for all SlurmProcess instances
     try:
-        _slurm_cluster_mediator = SlurmClusterMediator()
+        slurm_cluster_mediator = SlurmClusterMediator()
     except ValueError:
-        _slurm_cluster_mediator = None
+        slurm_cluster_mediator = None
         # we raise a ValueError if sacct/sinfo are not available
         logger.warning("Could not initialize SLURM cluster handling. "
                        "If you are sure SLURM (sinfo/sacct/etc) is available "
-                       "try calling `asyncmd.config.set_slurm_settings()` "
+                       "try calling `asyncmd.config.set_all_slurm_settings()` "
+                       "or `asyncmd.config.set_slurm_setting()` "
                        "with the appropriate arguments.")
     # we can not simply wait for the subprocess, since slurm exits directly
     # so we will sleep for this long between checks if slurm-job completed
@@ -79,12 +93,13 @@ class SlurmProcess:
     #       stderr and stdout
     sbatch_executable = "sbatch"
     scancel_executable = "scancel"
+    # default value for stdfile_removal
+    _stdfiles_removal = "success"
 
     def __init__(self, jobname: str, sbatch_script: str, *,
                  workdir: str | None = None,
                  time: float | None = None,
                  sbatch_options: dict | None = None,
-                 stdfiles_removal: str = "success",
                  **kwargs) -> None:
         """
         Initialize a `SlurmProcess`.
@@ -113,14 +128,6 @@ class SlurmProcess:
             ``--contiguous`` the dictionary needs to be ``{"contiguous": ""}``.
             See the SLURM documentation for a full list of sbatch options
             (https://slurm.schedmd.com/sbatch.html).
-        stdfiles_removal : str
-            Whether to remove the stdout, stderr (and possibly stdin) files.
-            Possible values are:
-
-             - "success": remove on successful completion, i.e. zero returncode)
-             - "no": never remove
-             - "yes"/"always": remove on job completion independent of
-               returncode and also when using :meth:`terminate`
 
         Raises
         ------
@@ -128,6 +135,13 @@ class SlurmProcess:
             If the value set via init kwarg for a attribute does not match the
             default/original type for that attribute.
         """
+        if not isinstance(self.slurm_cluster_mediator, SlurmClusterMediator):
+            raise SlurmError(
+                "SLURM monitoring not initialized. Please call "
+                "`asyncmd.config.set_slurm_setting()` or "
+                "`asyncmd.config.set_all_slurm_settings` with appropriate "
+                "arguments to ensure `sinfo` and `sacct` executables are available."
[…]
