### 2025-07-27 3a92ba2
Minor improvements for slurm module
