### 2025-07-27 cb2b5d4
tighten pylint configuration
