### 2025-07-29 b541669
Improved handling of h5py trajectory function value caches
