+++ src/asyncmd/trajectory/trajectory_cache.py
@@ -21,13 +21,18 @@ import io
 import os
 import abc
 import logging
+import typing
 import zipfile
 import collections
 
 import numpy as np
 
 
-from .._config import _GLOBALS
+from .._config import _GLOBALS, _GLOBALS_KEYS
+
+
+if typing.TYPE_CHECKING:  # pragma: no cover
+    import h5py
 
 
 logger = logging.getLogger(__name__)
@@ -37,6 +42,21 @@ class ValuesAlreadyStoredError(ValueError):
     """
     Error raised by :class:`TrajectoryFunctionValueCache` classes when trying to
     append values for a func_id that is already present.
+    """
+
+
+class NoValuesCachedForTrajectoryInH5PYError(ValueError):
+    """
+    Error raised by :class:`OneH5PYGroupTrajectoryFunctionValueCache` when
+    instantiating with a read-only h5py_cache in which no values for the given
+    :class:`Trajectory` (hash) are stored.
+    """
+
+
+class CanNotChangeReadOnlyH5PYError(PermissionError):
+    """
+    Error raised by :class:`OneH5PYGroupTrajectoryFunctionValueCache` when trying
+    to change (append or clear) an instance with a read-only h5py_cache.
     """
 
 
@@ -287,13 +307,20 @@ class TrajectoryFunctionValueCacheInNPZ(TrajectoryFunctionValueCache):
             os.unlink(self.fname_npz)  # and remove the file if it exists
 
 
-class TrajectoryFunctionValueCacheInH5PY(TrajectoryFunctionValueCache):
+class OneH5PYGroupTrajectoryFunctionValueCache(TrajectoryFunctionValueCache):
     """
-    Interface for caching trajectory function values in a given h5py group.
+    Interface for caching trajectory function values in one given h5py group, used
+    as composite class in :class:`TrajectoryFunctionValueCacheInH5PY`.
 
-    Gets the centrally set ``H5PY_CACHE`` configuration variable in ``__init__``
-    and uses it as ``h5py.Group`` to store the cached values in.
-    The values will be stored in this group in a subgroup (defined by
+    Can also be used on h5py files/groups open in read-only mode, then appending
+    and clearing will raise errors.
+
+    This class also contains a classmethod to copy over all cached values from
+    a given h5py (cache) group to another h5py (cache) group. This method is
+    (optionally) used when a new h5py cache is registered to transfer all values
+    from all previously registered caches.
+
+    The values will be stored in the given group in a subgroup (defined by
     ``self._H5PY_PATHS['prefix']``) and then by trajectory hash, i.e. the full path
     becomes '$self._H5PY_PATHS['prefix']/$TRAJ_HASH/'.
     Within this path/group there are two subgroups, one for the func_ids and one
@@ -309,19 +336,31 @@ class TrajectoryFunctionValueCacheInH5PY(TrajectoryFunctionValueCache):
                    "prefix": "asyncmd/TrajectoryFunctionValueCache",
                    }
 
-    def __init__(self, traj_hash: int, traj_files: list[str]) -> None:
+    def __init__(self, traj_hash: int, traj_files: list[str],
+                 h5py_cache: "h5py.Group | h5py.File",
+                 ) -> None:
         super().__init__(traj_hash, traj_files)
+        self.h5py_cache = h5py_cache
+        self._read_only = h5py_cache.file.mode == "r"
         try:
-            self._h5py_cache = _GLOBALS["H5PY_CACHE"]
-        except KeyError as e:
-            raise RuntimeError(
-                f"Can not initialize a {type(self)} without global h5py cache set!"
-                " Try calling `asyncmd.config.register_h5py_cache` first."
-                ) from e
-
-        self._root_grp = self._h5py_cache.require_group(
+            self._root_grp = self.h5py_cache.require_group(
[…]
