+++ src/asyncmd/config.py
@@ -22,12 +22,24 @@ import os
 import asyncio
 import logging
 import resource
+import typing
 
 
-from ._config import _GLOBALS, _SEMAPHORES, _OPT_SEMAPHORES
-from .trajectory.trajectory import _update_cache_type_for_all_trajectories
+from ._config import (_GLOBALS, _SEMAPHORES, _OPT_SEMAPHORES,
+                      _GLOBALS_KEYS, _SEMAPHORES_KEYS, _OPT_SEMAPHORES_KEYS,
+                      )
+from .trajectory.trajectory import (_update_cache_type_for_all_trajectories,
+                                    _deregister_h5py_cache_for_all_trajectories,
+                                    )
+from .trajectory.trajectory_cache import (
+        TrajectoryFunctionValueCacheInH5PY as _TrajectoryFunctionValueCacheInH5PY,
+        )
 # pylint: disable-next=unused-import
 from .slurm import set_slurm_settings, set_all_slurm_settings
+
+
+if typing.TYPE_CHECKING:  # pragma: no cover
+    import h5py
 
 
 logger = logging.getLogger(__name__)
@@ -62,7 +74,7 @@ def set_max_process(num: int | None = None, max_num: int | None = None) -> None:
             num = 2
     if max_num is not None:
         num = min((num, max_num))
-    _SEMAPHORES["MAX_PROCESS"] = asyncio.BoundedSemaphore(num)
+    _SEMAPHORES[_SEMAPHORES_KEYS.MAX_PROCESS] = asyncio.BoundedSemaphore(num)
 
 
 set_max_process()
@@ -121,7 +133,7 @@ def set_max_files_open(num: int | None = None, margin: int = 30) -> None:
     #       threads. The problem is that this semaphore will never be freed
     #       without any process getting a semaphore...
     semaval = int((num - margin) / 3)
-    _SEMAPHORES["MAX_FILES_OPEN"] = asyncio.BoundedSemaphore(semaval)
+    _SEMAPHORES[_SEMAPHORES_KEYS.MAX_FILES_OPEN] = asyncio.BoundedSemaphore(semaval)
 
 
 set_max_files_open()
@@ -147,9 +159,9 @@ def set_slurm_max_jobs(num: int | None) -> None:
     # pylint: disable-next=global-variable-not-assigned
     global _OPT_SEMAPHORES
     if num is None:
-        _OPT_SEMAPHORES["SLURM_MAX_JOB"] = None
+        _OPT_SEMAPHORES[_OPT_SEMAPHORES_KEYS.SLURM_MAX_JOB] = None
     else:
-        _OPT_SEMAPHORES["SLURM_MAX_JOB"] = asyncio.BoundedSemaphore(num)
+        _OPT_SEMAPHORES[_OPT_SEMAPHORES_KEYS.SLURM_MAX_JOB] = asyncio.BoundedSemaphore(num)
 
 
 set_slurm_max_jobs(num=None)
@@ -188,9 +200,9 @@ def set_trajectory_cache_type(cache_type: str,
     if (cache_type := cache_type.lower()) not in allowed_values:
         raise ValueError(f"Given cache type must be one of {allowed_values}."
                          + f" Was: {cache_type}.")
-    if _GLOBALS.get("TRAJECTORY_FUNCTION_CACHE_TYPE", "not_set") != cache_type:
+    if _GLOBALS.get(_GLOBALS_KEYS.TRAJECTORY_FUNCTION_CACHE_TYPE, "not_set") != cache_type:
         # only do something if the new cache type differs from what we have
-        _GLOBALS["TRAJECTORY_FUNCTION_CACHE_TYPE"] = cache_type
+        _GLOBALS[_GLOBALS_KEYS.TRAJECTORY_FUNCTION_CACHE_TYPE] = cache_type
[…]
