### 2025-07-29 86e231f
Use the newly introduced named-tuple for accessing _GLOBALS, _SEMAPHORES, and _OPT_SEMAPHORES everywhere
