### 2025-07-29 e8f3b6f
Improve documentation: Use intersphinx for h5py-references and explain the h5py cahching mechanism a bit more. Also make sure some alias/imported functions are properly crosslinked in the docs
