### 2025-07-29 c93063b
Add documentation on utility functions
