### 2025-07-29 b727022
make example notebooks more visible in documentation
