+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "asyncmd"
-version = "0.4.0"
+version = "0.4.1"
 dependencies = ["aiofiles",
                 "mdanalysis",
                 "numpy",
