### 2025-08-01 #8321 @espottesmith
@hejung I noticed that your Zenodo archive had a new version - 0.4.1. Should I update the version of record to this version?
