### 2025-08-01 #8321 @hejung
@espottesmith

> [@hejung](https://github.com/hejung) I noticed that your Zenodo archive had a new version - 0.4.1. Should I update the version of record to this version?

If that does not create  too much inconvenience it would be great! These were not things that came up in the review directly but rather things I noticed would be nice while implementing the review.

Thanks for the positive feedback!
