+++ joss_paper/paper.bib
@@ -95,7 +95,7 @@ year = {2011}
   pages = {7762--7774},
   publisher = {American Institute of Physics},
   issn = {0021-9606},
-  doi = {10/dg9x4z},
+  doi = {10.1063/1.1562614},
   urldate = {2020-04-11},
   abstract = {We derive a novel efficient scheme to measure the rate constant of transitions between stable states separated by high free energy barriers in a complex environment within the framework of transition path sampling. The method is based on directly and simultaneously measuring the fluxes through many phase space interfaces and increases the efficiency with at least a factor of 2 with respect to existing transition path sampling rate constant algorithms. The new algorithm is illustrated on the isomerization of a diatomic molecule immersed in a simple fluid.},
   langid = {english}
