### 2025-08-08 6656cfa
make MDEngine api more consistent
