### 2025-08-09 31e4ef7
gromacs.mdengine: improve apply_constraints and generate_velocities methods
