### 2025-08-09 45d5bf8
fix rare condition that resulted in the same TrajectoryFunctionWrapper being applied to the same Trajectory twice
