### 2025-08-09 d80ed49
make functions to update the cahche of all Trajectories 'public', i.e. remove their _ prefix
