### 2025-08-13 #8321 @Kevin-Mattheus-Moerman
- [ ] Please edit the archive title to match the paper title. You should be able to manually do this over at ZENODO.
- [ ] Please capitalise the `J` in `jupyter` so it reads `Jupyter`
