+++ joss_paper/paper.md
@@ -48,7 +48,7 @@ All computationally costly operations are submitted via the queuing system, i.e.
 
 `asyncmd` can therefore be used to manage large scale MD simulation campaigns on HPC resources. In addition, `asyncmd` provides an ideal building block to develop and implement trajectory based enhanced sampling methods as, e.g., variants of the string method [@e2002StringMethod], highly parallelized transition path or transition interface sampling methods [@dellago2002TransitionPathSampling;@vanerp2003NovelPathSampling], flux sampling methods [@chandler1978;@ruizmontero1997], or the weighted ensemble method [@huber1996Weightedensemble]. Showcasing its potential, `asyncmd` was already used to develop and implement the AI for molecular mechanism discovery (aimmd) algorithm [@jung2023;@aimmd], which adaptively steers a large number of simultaneous MD simulations using a combination of transition path sampling and machine learning.
 
-In addition to its documentation, `asyncmd` includes a number of examples in the form of jupyter notebooks. These notebooks illustrate most of the common operations, such as performing MD simulations (possibly until a condition is fulfilled) or how to extract configurations from trajectories to restart MD simulations from it, but also include an example showcasing how to implement the weighted ensemble method.
+In addition to its documentation, `asyncmd` includes a number of examples in the form of Jupyter notebooks. These notebooks illustrate most of the common operations, such as performing MD simulations (possibly until a condition is fulfilled) or how to extract configurations from trajectories to restart MD simulations from it, but also include an example showcasing how to implement the weighted ensemble method.
 
 ## State of the field
 
