### 2025-08-14 #8321 @hejung
@Kevin-Mattheus-Moerman Thank you for handling the final stage of the submission!

I just now updated the paper branch (fixing the typo) and the Zenodo now also has the same title as the publication. This should address your remaining points, but please let me know if there is more to do from our side!
