### 2025-08-19 #8321 @hejung
I just wanted to thank everyone involved in the very enjoyable and transparent review process! It really improved the quality of asyncmd, its documentation, and the manuscript 🎉🎉🎉
